refactor: decompose ConfirmationShell from PermissionHandler
