+++ packages/gator-permissions-snap/src/core/confirmation/constants.ts
@@ -1,0 +1,1 @@
+export const JUSTIFICATION_SHOW_MORE_BUTTON_NAME = 'show-more-justification';
