+++ packages/gator-permissions-snap/test/core/confirmation/ConfirmationShell.test.ts
@@ -1,0 +1,3786 @@
+import { describe, expect, it, jest } from '@jest/globals';
+import type { PermissionRequest } from '@metamask/7715-permissions-shared/types';
+import { NO_ASSET_ADDRESS } from '@metamask/7715-permissions-shared/types';
+import { UserInputEventType } from '@metamask/snaps-sdk';
+import { Text } from '@metamask/snaps-sdk/jsx';
+import type { SnapElement } from '@metamask/snaps-sdk/jsx';
+
+import { AddressScanResultType } from '../../../src/clients/trustSignalsClient';
+import type { TokenBalanceAndMetadata } from '../../../src/clients/types';
+import type { AccountController } from '../../../src/core/accountController';
+import { ConfirmationShell } from '../../../src/core/confirmation/ConfirmationShell';
+import { ExistingPermissionsState } from '../../../src/core/existingpermissions/existingPermissionsState';
+import { METAMASK_FACILITATOR_ADDRESSES } from '../../../src/core/facilitatorAddresses';
+import type { BaseContext, RuleDefinition } from '../../../src/core/types';
+import type { TokenMetadataService } from '../../../src/services/tokenMetadataService';
+import type { TokenPricesService } from '../../../src/services/tokenPricesService';
+import type {
+  UserEventDispatcher,
+  UserEventHandler,
+} from '../../../src/userEventDispatcher';
+import type { MessageKey } from '../../../src/utils/i18n';
+
+const mockAddress = '0x1234567890123456789012345678901234567890' as const;
+const mockAddress2 = '0x1234567890123456789012345678901234567891' as const;
+const mockAssetAddress = '0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3' as const;
+const mockInterfaceId = 'test-interface-id';
+const mockOrigin = 'https://example.com';
+const mockTokenBalanceFiat = '$1000';
+
+type TestContextType = BaseContext;
+type TestMetadataType = object;
+
+const mockPermissionRequest: PermissionRequest = {
+  chainId: '0x1',
+  to: mockAddress,
+  permission: {
+    type: 'native-token-stream',
+    data: {
+      amountPerSecond: '0x1',
+      startTime: 1234567890,
+      justification: 'test',
+    },
+    isAdjustmentAllowed: false,
+  },
+  rules: [],
+};
+
+const mockContext: TestContextType = {
+  justification:
+    'Test justification text that is longer than twenty characters',
+  tokenMetadata: {
+    symbol: 'ETH',
+    decimals: 18,
+    iconDataBase64: null,
+  },
+  accountAddressCaip10: `eip155:1:${mockAddress}`,
+  tokenAddressCaip19: `eip155:1/erc20:${mockAssetAddress}`,
+  expiry: {
+    timestamp: 1234567890,
+  },
+  isAdjustmentAllowed: false,
+};
+
+const mockTokenBalanceAndMetadata: TokenBalanceAndMetadata = {
+  balance: 1000000000000000000n,
+  symbol: 'ETH',
+  decimals: 18,
+};
+const mockMetadata: TestMetadataType = {};
+
+const mockBodyContent = Text({
+  children: 'Permission body',
+}) as unknown as SnapElement;
+
+// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
+const setupTest = (options?: { rules?: RuleDefinition<any, any>[] }) => {
+  const title = 'permissionRequestTitle' as MessageKey;
+  const subtitle = 'permissionRequestSubtitle' as MessageKey;
+  const rules = options?.rules ?? [];
+
+  const boundEvents = new Map<string, UserEventHandler<UserInputEventType>>();
+  const unboundEvents = new Map<string, UserEventHandler<UserInputEventType>>();
+
+  // eslint-disable-next-line prefer-const
+  let userEventDispatcher: jest.Mocked<UserEventDispatcher>;
+  const bindEvent = ({
+    elementName,
+    eventType,
+    interfaceId,
+    handler,
+  }: {
+    elementName: string;
+    eventType: string;
+    interfaceId: string;
+    handler: UserEventHandler<UserInputEventType>;
+  }): { unbind: () => void; dispatcher: jest.Mocked<UserEventDispatcher> } => {
+    boundEvents.set(`${elementName}:${eventType}:${interfaceId}`, handler);
+
+    return {
+      unbind: (): void => {
+        unboundEvents.set(
+          `${elementName}:${eventType}:${interfaceId}`,
+          handler,
+        );
+      },
+      dispatcher: userEventDispatcher,
+    };
+  };
+
+  const getBoundEvent = (args: {
+    elementName: string;
+    eventType: string;
+    interfaceId: string;
+  }): UserEventHandler<UserInputEventType> | undefined => {
+    return boundEvents.get(
+      `${args.elementName}:${args.eventType}:${args.interfaceId}`,
+    );
+  };
+
+  const getUnboundEvent = (args: {
+    elementName: string;
+    eventType: string;
+    interfaceId: string;
+  }): UserEventHandler<UserInputEventType> | undefined => {
+    return unboundEvents.get(
+      `${args.elementName}:${args.eventType}:${args.interfaceId}`,
+    );
+  };
+
+  const accountController = {
+    signDelegation: jest.fn(),
+    getAccountAddresses: jest.fn(),
+    getAccountUpgradeStatus: jest.fn(),
+    upgradeAccount: jest.fn(),
+  } as unknown as jest.Mocked<AccountController>;
+
+  accountController.getAccountUpgradeStatus.mockResolvedValue({
+    isUpgraded: false,
+  });
+
+  userEventDispatcher = {
+    on: jest.fn(bindEvent),
+    off: jest.fn(),
+    createUserInputEventHandler: jest.fn(),
+    waitForPendingHandlers: jest.fn(),
+  } as unknown as jest.Mocked<UserEventDispatcher>;
+
+  const tokenPricesService = {
+    getCryptoToFiatConversion: jest.fn(async () =>
+      Promise.resolve(mockTokenBalanceFiat),
+    ),
+  } as unknown as jest.Mocked<TokenPricesService>;
+
+  const tokenMetadataService = {
+    getTokenBalanceAndMetadata: jest.fn(async () =>
+      Promise.resolve(mockTokenBalanceAndMetadata),
+    ),
+    fetchIconDataAsBase64: jest.fn(),
+  } as unknown as jest.Mocked<TokenMetadataService>;
+
+  const renderBody = jest.fn(async () => Promise.resolve(mockBodyContent));
+
+  const confirmationShell = new ConfirmationShell({
+    userEventDispatcher,
+    accountController,
+    tokenMetadataService,
+    tokenPricesService,
+    title,
+    subtitle,
+    permissionRequest: mockPermissionRequest,
+    renderBody,
+  });
+
+  const updateContext =
+    jest.fn<(args: { updatedContext: TestContextType }) => Promise<void>>();
+
+  return {
+    confirmationShell,
+    renderBody,
+    updateContext,
+    rules,
+    getBoundEvent,
+    getUnboundEvent,
+    tokenMetadataService,
+    tokenPricesService,
+    accountController,
+    userEventDispatcher,
+  };
+};
+
+describe('ConfirmationShell', () => {
+  describe('createSkeletonContent', () => {
+    it('creates skeleton confirmation content', () => {
+      const { confirmationShell } = setupTest();
+
+      const result = confirmationShell.createSkeletonContent();
+
+      expect(result).toBeDefined();
+      expect(result.type).toBe('Container');
+    });
+  });
+
+  describe('createConfirmationContent', () => {
+    it('calls createConfirmationContent to produce the permission specific content', async () => {
+      const { confirmationShell, renderBody } = setupTest();
+      await confirmationShell.createConfirmationContent({
+        context: mockContext,
+        metadata: mockMetadata,
+        origin: mockOrigin,
+        chainId: 1,
+        scanDappUrlResult: null,
+        scanAddressResult: null,
+        existingPermissionsStatus: ExistingPermissionsState.None,
+        isGrantDisabled: false,
+      });
+
+      expect(renderBody).toHaveBeenCalledWith({
+        context: mockContext,
+        metadata: mockMetadata,
+      });
+    });
+
+    it('uses translated fallback for address warning when scanAddressResult.label is empty', async () => {
+      const { confirmationShell } = setupTest();
+      const result = await confirmationShell.createConfirmationContent({
+        context: mockContext,
+        metadata: mockMetadata,
+        origin: mockOrigin,
+        chainId: 1,
+        scanDappUrlResult: null,
+        scanAddressResult: {
+          resultType: AddressScanResultType.Malicious,
+          label: '',
+        },
+        existingPermissionsStatus: ExistingPermissionsState.None,
+        isGrantDisabled: false,
+      });
+
+      // When label is empty, confirmationShellContent should use t('maliciousAddressLabel')
+      const serialized = JSON.stringify(result);
+      expect(serialized).toContain('Malicious address');
+    });
+
+    it('renders redeemer addresses as redeemers when no payee rule is present', async () => {
+      const { confirmationShell } = setupTest();
+      const result = await confirmationShell.createConfirmationContent({
+        context: {
+          ...mockContext,
+          redeemerAddresses: ['0x1111111111111111111111111111111111111111'],
+        },
+        metadata: mockMetadata,
+        origin: mockOrigin,
+        chainId: 1,
+        scanDappUrlResult: null,
+        scanAddressResult: null,
+        existingPermissionsStatus: ExistingPermissionsState.None,
+        isGrantDisabled: false,
+      });
+
+      const serialized = JSON.stringify(result);
+      expect(serialized).toContain('Redeemers');
+      expect(serialized).not.toContain('Facilitators');
+    });
+
+    it('renders known MetaMask facilitator redeemer addresses as a MetaMask facilitator redeemer', async () => {
+      const { confirmationShell } = setupTest();
+      const result = await confirmationShell.createConfirmationContent({
+        context: {
+          ...mockContext,
+          redeemerAddresses: [...METAMASK_FACILITATOR_ADDRESSES],
+        },
+        metadata: mockMetadata,
+        origin: mockOrigin,
+        chainId: 1,
+        scanDappUrlResult: null,
+        scanAddressResult: null,
+        existingPermissionsStatus: ExistingPermissionsState.None,
+        isGrantDisabled: false,
+      });
+
+      const serialized = JSON.stringify(result);
+      expect(serialized).toContain('Redeemers');
+      expect(serialized).toContain('MetaMask facilitator');
+      expect(serialized).toContain(
+        'Only these addresses may redeem this permission.',
+      );
+      expect(serialized).not.toContain('Facilitators');
+    });
+
+    it('does not render arbitrary redeemer addresses as facilitators when a payee rule is present', async () => {
+      const { confirmationShell } = setupTest();
+      const result = await confirmationShell.createConfirmationContent({
+        context: {
+          ...mockContext,
+          redeemerAddresses: ['0x1111111111111111111111111111111111111111'],
+          payeeAddresses: ['0x2222222222222222222222222222222222222222'],
+        },
+        metadata: mockMetadata,
+        origin: mockOrigin,
+        chainId: 1,
+        scanDappUrlResult: null,
+        scanAddressResult: null,
+        existingPermissionsStatus: ExistingPermissionsState.None,
+        isGrantDisabled: false,
+      });
+
+      const serialized = JSON.stringify(result);
+      expect(serialized).toContain('Redeemers');
+      expect(serialized).not.toContain('Facilitators');
+    });
+  });
+
+  describe('bindSessionEvents', () => {
+    it('registers event handlers for account selection and justification toggle', async () => {
+      const rule: RuleDefinition<TestContextType, TestMetadataType> = {
+        name: 'amountPerSecond',
+        label: 'amountLabel',
+        type: 'number',
+        getRuleData: () => ({
+          value: '0x1',
+          isVisible: true,
+          isEditable: false,
+        }),
+        updateContext: (context) => context,
+      };
+      const { confirmationShell, rules, getBoundEvent, updateContext } =
+        setupTest({ rules: [rule] });
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: mockContext,
+        rules,
+        updateContext,
+      });
+
+      const accountSelectorBoundEvent = getBoundEvent({
+        elementName: 'account-selector',
+        eventType: 'InputChangeEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      const showMoreButtonBoundEvent = getBoundEvent({
+        elementName: 'show-more-justification',
+        eventType: 'ButtonClickEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(accountSelectorBoundEvent).toBeDefined();
+      expect(showMoreButtonBoundEvent).toBeDefined();
+    });
+
+    it('loads the balance for the selected account', async () => {
+      const { confirmationShell, rules, tokenMetadataService, updateContext } =
+        setupTest();
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: mockContext,
+        rules,
+        updateContext,
+      });
+
+      expect(
+        tokenMetadataService.getTokenBalanceAndMetadata,
+      ).toHaveBeenCalledWith({
+        chainId: 1,
+        account: mockAddress,
+        assetAddress: mockAssetAddress,
+      });
+    });
+
+    it('updates the context when the account is changed', async () => {
+      const { confirmationShell, rules, getBoundEvent, updateContext } =
+        setupTest();
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: mockContext,
+        rules,
+        updateContext,
+      });
+
+      const accountSelectorChangeHandler = getBoundEvent({
+        elementName: 'account-selector',
+        eventType: 'InputChangeEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(accountSelectorChangeHandler).toBeDefined();
+
+      const mockAddress2Caip10 = `eip155:1:${mockAddress2}`;
+
+      await accountSelectorChangeHandler?.({
+        event: {
+          value: { addresses: [mockAddress2Caip10] } as any,
+          name: 'account-selector',
+          type: UserInputEventType.InputChangeEvent,
+        },
+        interfaceId: mockInterfaceId,
+      });
+
+      const expectedUpdatedContext = {
+        ...mockContext,
+        accountAddressCaip10: mockAddress2Caip10,
+      };
+
+      expect(updateContext).toHaveBeenCalledWith({
+        updatedContext: expectedUpdatedContext,
+      });
+    });
+
+    it('updates the balance when the account is changed', async () => {
+      const {
+        confirmationShell,
+        rules,
+        getBoundEvent,
+        tokenMetadataService,
+        updateContext,
+      } = setupTest();
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: mockContext,
+        rules,
+        updateContext,
+      });
+
+      const accountSelectorChangeHandler = getBoundEvent({
+        elementName: 'account-selector',
+        eventType: 'InputChangeEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(accountSelectorChangeHandler).toBeDefined();
+
+      const mockAddress2Caip10 = `eip155:1:${mockAddress2}`;
+
+      await accountSelectorChangeHandler?.({
+        event: {
+          value: { addresses: [mockAddress2Caip10] } as any,
+          name: 'account-selector',
+          type: UserInputEventType.InputChangeEvent,
+        },
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(
+        tokenMetadataService.getTokenBalanceAndMetadata,
+      ).toHaveBeenCalledTimes(2);
+
+      expect(
+        tokenMetadataService.getTokenBalanceAndMetadata,
+      ).toHaveBeenCalledWith({
+        chainId: 1,
+        account: mockAddress2,
+        assetAddress: mockAssetAddress,
+      });
+    });
+
+    it('renders the balance in the confirmation content', async () => {
+      const { confirmationShell, rules, updateContext } = setupTest();
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: mockContext,
+        rules,
+        updateContext,
+      });
+
+      const confirmationContent =
+        await confirmationShell.createConfirmationContent({
+          context: mockContext,
+          metadata: mockMetadata,
+          origin: mockOrigin,
+          chainId: 1,
+          scanDappUrlResult: null,
+          scanAddressResult: null,
+          existingPermissionsStatus: ExistingPermissionsState.None,
+          isGrantDisabled: false,
+        });
+      expect(confirmationContent).toMatchInlineSnapshot(`
+{
+  "key": null,
+  "props": {
+    "children": [
+      {
+        "key": null,
+        "props": {
+          "children": {
+            "key": null,
+            "props": {
+              "children": [
+                {
+                  "key": null,
+                  "props": {
+                    "center": true,
+                    "children": [
+                      {
+                        "key": null,
+                        "props": {
+                          "children": "Permission request",
+                          "size": "lg",
+                        },
+                        "type": "Heading",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "children": "This site wants permissions to spend your tokens.",
+                        },
+                        "type": "Text",
+                      },
+                    ],
+                  },
+                  "type": "Box",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": {
+                      "key": null,
+                      "props": {
+                        "children": [
+                          {
+                            "key": null,
+                            "props": {
+                              "alignment": "space-between",
+                              "children": {
+                                "key": null,
+                                "props": {
+                                  "children": [
+                                    {
+                                      "key": null,
+                                      "props": {
+                                        "children": "Account",
+                                      },
+                                      "type": "Text",
+                                    },
+                                    {
+                                      "key": null,
+                                      "props": {
+                                        "children": {
+                                          "key": null,
+                                          "props": {
+                                            "color": "muted",
+                                            "name": "question",
+                                            "size": "inherit",
+                                          },
+                                          "type": "Icon",
+                                        },
+                                        "content": {
+                                          "key": null,
+                                          "props": {
+                                            "children": "The account from which the permission is being granted.",
+                                          },
+                                          "type": "Text",
+                                        },
+                                      },
+                                      "type": "Tooltip",
+                                    },
+                                  ],
+                                  "direction": "horizontal",
+                                },
+                                "type": "Box",
+                              },
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "chainIds": [
+                                "eip155:1",
+                              ],
+                              "name": "account-selector",
+                              "switchGlobalAccount": false,
+                              "value": "eip155:1:0x1234567890123456789012345678901234567890",
+                            },
+                            "type": "AccountSelector",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "children": "This account will be upgraded to a smart account to complete this permission.",
+                              "color": "warning",
+                              "size": "sm",
+                            },
+                            "type": "Text",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "alignment": "end",
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {},
+                                  "type": "Skeleton",
+                                },
+                                {
+                                  "key": null,
+                                  "props": {},
+                                  "type": "Skeleton",
+                                },
+                              ],
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                        ],
+                        "direction": "vertical",
+                      },
+                      "type": "Box",
+                    },
+                  },
+                  "type": "Section",
+                },
+                false,
+                false,
+                {
+                  "key": null,
+                  "props": {
+                    "children": {
+                      "key": null,
+                      "props": {
+                        "alignment": "space-between",
+                        "children": [
+                          {
+                            "key": null,
+                            "props": {
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "Justification",
+                                  },
+                                  "type": "Text",
+                                },
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": {
+                                      "key": null,
+                                      "props": {
+                                        "color": "muted",
+                                        "name": "question",
+                                        "size": "inherit",
+                                      },
+                                      "type": "Icon",
+                                    },
+                                    "content": {
+                                      "key": null,
+                                      "props": {
+                                        "children": "Justification given by the recipient for requesting this permission.",
+                                      },
+                                      "type": "Text",
+                                    },
+                                  },
+                                  "type": "Tooltip",
+                                },
+                              ],
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "Test justification text that is longer than twenty characters",
+                                  },
+                                  "type": "Text",
+                                },
+                                null,
+                              ],
+                              "direction": "vertical",
+                            },
+                            "type": "Box",
+                          },
+                        ],
+                        "direction": "vertical",
+                      },
+                      "type": "Box",
+                    },
+                  },
+                  "type": "Section",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": [
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Request from",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The site requesting the permission",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "alignment": "end",
+                                      "children": "https://example.com",
+                                    },
+                                    "type": "Text",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Recipient",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The site requesting the permission",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": {
+                                        "key": null,
+                                        "props": {
+                                          "children": "0x12345...67890",
+                                        },
+                                        "type": "Text",
+                                      },
+                                      "content": "0x1234567890123456789012345678901234567890",
+                                    },
+                                    "type": "Tooltip",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Network",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The network on which the permission is being requested",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "alignment": "end",
+                                      "children": "Ethereum Mainnet",
+                                    },
+                                    "type": "Text",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Token",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The token being requested",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": {
+                                        "key": null,
+                                        "props": {
+                                          "children": "ETH",
+                                          "href": "https://etherscan.io/address/0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3",
+                                        },
+                                        "type": "Link",
+                                      },
+                                      "content": "0x38c4A...611F3",
+                                    },
+                                    "type": "Tooltip",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      null,
+                      null,
+                    ],
+                  },
+                  "type": "Section",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": "Permission body",
+                  },
+                  "type": "Text",
+                },
+              ],
+              "direction": "vertical",
+            },
+            "type": "Box",
+          },
+        },
+        "type": "Box",
+      },
+      {
+        "key": null,
+        "props": {
+          "children": [
+            {
+              "key": null,
+              "props": {
+                "children": "Cancel",
+                "name": "cancel-button",
+                "variant": "destructive",
+              },
+              "type": "Button",
+            },
+            {
+              "key": null,
+              "props": {
+                "children": "Grant",
+                "disabled": false,
+                "name": "grant-button",
+                "variant": "primary",
+              },
+              "type": "Button",
+            },
+          ],
+        },
+        "type": "Footer",
+      },
+    ],
+  },
+  "type": "Container",
+}
+`);
+    });
+
+    it('updates the balance in the confirmation content when the account is changed', async () => {
+      const {
+        confirmationShell,
+        rules,
+        getBoundEvent,
+        tokenPricesService,
+        updateContext,
+      } = setupTest();
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: mockContext,
+        rules,
+        updateContext,
+      });
+
+      const accountSelectorChangeHandler = getBoundEvent({
+        elementName: 'account-selector',
+        eventType: 'InputChangeEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(accountSelectorChangeHandler).toBeDefined();
+
+      const mockAddress2Caip10 = `eip155:1:${mockAddress2}`;
+
+      tokenPricesService.getCryptoToFiatConversion.mockResolvedValue('$2000');
+
+      await accountSelectorChangeHandler?.({
+        event: {
+          value: { addresses: [mockAddress2Caip10] } as any,
+          name: 'account-selector',
+          type: UserInputEventType.InputChangeEvent,
+        },
+        interfaceId: mockInterfaceId,
+      });
+
+      const confirmationContent =
+        await confirmationShell.createConfirmationContent({
+          context: mockContext,
+          metadata: mockMetadata,
+          origin: mockOrigin,
+          chainId: 1,
+          scanDappUrlResult: null,
+          scanAddressResult: null,
+          existingPermissionsStatus: ExistingPermissionsState.None,
+          isGrantDisabled: false,
+        });
+
+      expect(confirmationContent).toMatchInlineSnapshot(`
+{
+  "key": null,
+  "props": {
+    "children": [
+      {
+        "key": null,
+        "props": {
+          "children": {
+            "key": null,
+            "props": {
+              "children": [
+                {
+                  "key": null,
+                  "props": {
+                    "center": true,
+                    "children": [
+                      {
+                        "key": null,
+                        "props": {
+                          "children": "Permission request",
+                          "size": "lg",
+                        },
+                        "type": "Heading",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "children": "This site wants permissions to spend your tokens.",
+                        },
+                        "type": "Text",
+                      },
+                    ],
+                  },
+                  "type": "Box",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": {
+                      "key": null,
+                      "props": {
+                        "children": [
+                          {
+                            "key": null,
+                            "props": {
+                              "alignment": "space-between",
+                              "children": {
+                                "key": null,
+                                "props": {
+                                  "children": [
+                                    {
+                                      "key": null,
+                                      "props": {
+                                        "children": "Account",
+                                      },
+                                      "type": "Text",
+                                    },
+                                    {
+                                      "key": null,
+                                      "props": {
+                                        "children": {
+                                          "key": null,
+                                          "props": {
+                                            "color": "muted",
+                                            "name": "question",
+                                            "size": "inherit",
+                                          },
+                                          "type": "Icon",
+                                        },
+                                        "content": {
+                                          "key": null,
+                                          "props": {
+                                            "children": "The account from which the permission is being granted.",
+                                          },
+                                          "type": "Text",
+                                        },
+                                      },
+                                      "type": "Tooltip",
+                                    },
+                                  ],
+                                  "direction": "horizontal",
+                                },
+                                "type": "Box",
+                              },
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "chainIds": [
+                                "eip155:1",
+                              ],
+                              "name": "account-selector",
+                              "switchGlobalAccount": false,
+                              "value": "eip155:1:0x1234567890123456789012345678901234567890",
+                            },
+                            "type": "AccountSelector",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "children": "This account will be upgraded to a smart account to complete this permission.",
+                              "color": "warning",
+                              "size": "sm",
+                            },
+                            "type": "Text",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "alignment": "end",
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {},
+                                  "type": "Skeleton",
+                                },
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": [
+                                      "1",
+                                      " ",
+                                      "available",
+                                    ],
+                                  },
+                                  "type": "Text",
+                                },
+                              ],
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                        ],
+                        "direction": "vertical",
+                      },
+                      "type": "Box",
+                    },
+                  },
+                  "type": "Section",
+                },
+                false,
+                false,
+                {
+                  "key": null,
+                  "props": {
+                    "children": {
+                      "key": null,
+                      "props": {
+                        "alignment": "space-between",
+                        "children": [
+                          {
+                            "key": null,
+                            "props": {
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "Justification",
+                                  },
+                                  "type": "Text",
+                                },
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": {
+                                      "key": null,
+                                      "props": {
+                                        "color": "muted",
+                                        "name": "question",
+                                        "size": "inherit",
+                                      },
+                                      "type": "Icon",
+                                    },
+                                    "content": {
+                                      "key": null,
+                                      "props": {
+                                        "children": "Justification given by the recipient for requesting this permission.",
+                                      },
+                                      "type": "Text",
+                                    },
+                                  },
+                                  "type": "Tooltip",
+                                },
+                              ],
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "Test justification text that is longer than twenty characters",
+                                  },
+                                  "type": "Text",
+                                },
+                                null,
+                              ],
+                              "direction": "vertical",
+                            },
+                            "type": "Box",
+                          },
+                        ],
+                        "direction": "vertical",
+                      },
+                      "type": "Box",
+                    },
+                  },
+                  "type": "Section",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": [
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Request from",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The site requesting the permission",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "alignment": "end",
+                                      "children": "https://example.com",
+                                    },
+                                    "type": "Text",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Recipient",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The site requesting the permission",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": {
+                                        "key": null,
+                                        "props": {
+                                          "children": "0x12345...67890",
+                                        },
+                                        "type": "Text",
+                                      },
+                                      "content": "0x1234567890123456789012345678901234567890",
+                                    },
+                                    "type": "Tooltip",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Network",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The network on which the permission is being requested",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "alignment": "end",
+                                      "children": "Ethereum Mainnet",
+                                    },
+                                    "type": "Text",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Token",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The token being requested",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": {
+                                        "key": null,
+                                        "props": {
+                                          "children": "ETH",
+                                          "href": "https://etherscan.io/address/0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3",
+                                        },
+                                        "type": "Link",
+                                      },
+                                      "content": "0x38c4A...611F3",
+                                    },
+                                    "type": "Tooltip",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      null,
+                      null,
+                    ],
+                  },
+                  "type": "Section",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": "Permission body",
+                  },
+                  "type": "Text",
+                },
+              ],
+              "direction": "vertical",
+            },
+            "type": "Box",
+          },
+        },
+        "type": "Box",
+      },
+      {
+        "key": null,
+        "props": {
+          "children": [
+            {
+              "key": null,
+              "props": {
+                "children": "Cancel",
+                "name": "cancel-button",
+                "variant": "destructive",
+              },
+              "type": "Button",
+            },
+            {
+              "key": null,
+              "props": {
+                "children": "Grant",
+                "disabled": false,
+                "name": "grant-button",
+                "variant": "primary",
+              },
+              "type": "Button",
+            },
+          ],
+        },
+        "type": "Footer",
+      },
+    ],
+  },
+  "type": "Container",
+}
+`);
+    });
+
+    it('renders skeletons while the balance is loading', async () => {
+      const {
+        confirmationShell,
+        rules,
+        getBoundEvent,
+        tokenMetadataService,
+        tokenPricesService,
+        updateContext,
+      } = setupTest();
+
+      let resolveTokenBalancePromise: () => void = (): void => {
+        throw new Error('Function should never be called');
+      };
+      const tokenBalancePromise = new Promise<TokenBalanceAndMetadata>(
+        (resolve) => {
+          resolveTokenBalancePromise = (): void =>
+            resolve(mockTokenBalanceAndMetadata);
+        },
+      );
+
+      tokenMetadataService.getTokenBalanceAndMetadata.mockReturnValue(
+        tokenBalancePromise,
+      );
+
+      let resolveFiatBalancePromise: () => void = (): void => {
+        throw new Error('Function should never be called');
+      };
+      const fiatBalancePromise = new Promise<string>((resolve) => {
+        resolveFiatBalancePromise = (): void => resolve(mockTokenBalanceFiat);
+      });
+      tokenPricesService.getCryptoToFiatConversion.mockReturnValue(
+        fiatBalancePromise,
+      );
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: mockContext,
+        rules,
+        updateContext,
+      });
+
+      const accountSelectorChangeHandler = getBoundEvent({
+        elementName: 'account-selector',
+        eventType: 'InputChangeEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(accountSelectorChangeHandler).toBeDefined();
+
+      const mockAddress2Caip10 = `eip155:1:${mockAddress2}`;
+
+      await accountSelectorChangeHandler?.({
+        event: {
+          value: { addresses: [mockAddress2Caip10] } as any,
+          name: 'account-selector',
+          type: UserInputEventType.InputChangeEvent,
+        },
+        interfaceId: mockInterfaceId,
+      });
+
+      // called twice:
+      // 1. with the context updated to include the new account address
+      // 2. after the account upgrade status is fetched for the new account
+      expect(updateContext).toHaveBeenCalledTimes(2);
+      expect(updateContext).toHaveBeenCalledWith({
+        updatedContext: {
+          ...mockContext,
+          accountAddressCaip10: mockAddress2Caip10,
+        },
+      });
+
+      const confirmationContent =
+        await confirmationShell.createConfirmationContent({
+          context: mockContext,
+          metadata: mockMetadata,
+          origin: mockOrigin,
+          chainId: 1,
+          scanDappUrlResult: null,
+          scanAddressResult: null,
+          existingPermissionsStatus: ExistingPermissionsState.None,
+          isGrantDisabled: false,
+        });
+
+      // skeletons in place of both the token balance and fiat balance
+      expect(confirmationContent).toMatchInlineSnapshot(`
+{
+  "key": null,
+  "props": {
+    "children": [
+      {
+        "key": null,
+        "props": {
+          "children": {
+            "key": null,
+            "props": {
+              "children": [
+                {
+                  "key": null,
+                  "props": {
+                    "center": true,
+                    "children": [
+                      {
+                        "key": null,
+                        "props": {
+                          "children": "Permission request",
+                          "size": "lg",
+                        },
+                        "type": "Heading",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "children": "This site wants permissions to spend your tokens.",
+                        },
+                        "type": "Text",
+                      },
+                    ],
+                  },
+                  "type": "Box",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": {
+                      "key": null,
+                      "props": {
+                        "children": [
+                          {
+                            "key": null,
+                            "props": {
+                              "alignment": "space-between",
+                              "children": {
+                                "key": null,
+                                "props": {
+                                  "children": [
+                                    {
+                                      "key": null,
+                                      "props": {
+                                        "children": "Account",
+                                      },
+                                      "type": "Text",
+                                    },
+                                    {
+                                      "key": null,
+                                      "props": {
+                                        "children": {
+                                          "key": null,
+                                          "props": {
+                                            "color": "muted",
+                                            "name": "question",
+                                            "size": "inherit",
+                                          },
+                                          "type": "Icon",
+                                        },
+                                        "content": {
+                                          "key": null,
+                                          "props": {
+                                            "children": "The account from which the permission is being granted.",
+                                          },
+                                          "type": "Text",
+                                        },
+                                      },
+                                      "type": "Tooltip",
+                                    },
+                                  ],
+                                  "direction": "horizontal",
+                                },
+                                "type": "Box",
+                              },
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "chainIds": [
+                                "eip155:1",
+                              ],
+                              "name": "account-selector",
+                              "switchGlobalAccount": false,
+                              "value": "eip155:1:0x1234567890123456789012345678901234567890",
+                            },
+                            "type": "AccountSelector",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "children": "This account will be upgraded to a smart account to complete this permission.",
+                              "color": "warning",
+                              "size": "sm",
+                            },
+                            "type": "Text",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "alignment": "end",
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {},
+                                  "type": "Skeleton",
+                                },
+                                {
+                                  "key": null,
+                                  "props": {},
+                                  "type": "Skeleton",
+                                },
+                              ],
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                        ],
+                        "direction": "vertical",
+                      },
+                      "type": "Box",
+                    },
+                  },
+                  "type": "Section",
+                },
+                false,
+                false,
+                {
+                  "key": null,
+                  "props": {
+                    "children": {
+                      "key": null,
+                      "props": {
+                        "alignment": "space-between",
+                        "children": [
+                          {
+                            "key": null,
+                            "props": {
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "Justification",
+                                  },
+                                  "type": "Text",
+                                },
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": {
+                                      "key": null,
+                                      "props": {
+                                        "color": "muted",
+                                        "name": "question",
+                                        "size": "inherit",
+                                      },
+                                      "type": "Icon",
+                                    },
+                                    "content": {
+                                      "key": null,
+                                      "props": {
+                                        "children": "Justification given by the recipient for requesting this permission.",
+                                      },
+                                      "type": "Text",
+                                    },
+                                  },
+                                  "type": "Tooltip",
+                                },
+                              ],
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "Test justification text that is longer than twenty characters",
+                                  },
+                                  "type": "Text",
+                                },
+                                null,
+                              ],
+                              "direction": "vertical",
+                            },
+                            "type": "Box",
+                          },
+                        ],
+                        "direction": "vertical",
+                      },
+                      "type": "Box",
+                    },
+                  },
+                  "type": "Section",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": [
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Request from",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The site requesting the permission",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "alignment": "end",
+                                      "children": "https://example.com",
+                                    },
+                                    "type": "Text",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Recipient",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The site requesting the permission",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": {
+                                        "key": null,
+                                        "props": {
+                                          "children": "0x12345...67890",
+                                        },
+                                        "type": "Text",
+                                      },
+                                      "content": "0x1234567890123456789012345678901234567890",
+                                    },
+                                    "type": "Tooltip",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Network",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The network on which the permission is being requested",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "alignment": "end",
+                                      "children": "Ethereum Mainnet",
+                                    },
+                                    "type": "Text",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Token",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The token being requested",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": {
+                                        "key": null,
+                                        "props": {
+                                          "children": "ETH",
+                                          "href": "https://etherscan.io/address/0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3",
+                                        },
+                                        "type": "Link",
+                                      },
+                                      "content": "0x38c4A...611F3",
+                                    },
+                                    "type": "Tooltip",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      null,
+                      null,
+                    ],
+                  },
+                  "type": "Section",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": "Permission body",
+                  },
+                  "type": "Text",
+                },
+              ],
+              "direction": "vertical",
+            },
+            "type": "Box",
+          },
+        },
+        "type": "Box",
+      },
+      {
+        "key": null,
+        "props": {
+          "children": [
+            {
+              "key": null,
+              "props": {
+                "children": "Cancel",
+                "name": "cancel-button",
+                "variant": "destructive",
+              },
+              "type": "Button",
+            },
+            {
+              "key": null,
+              "props": {
+                "children": "Grant",
+                "disabled": false,
+                "name": "grant-button",
+                "variant": "primary",
+              },
+              "type": "Button",
+            },
+          ],
+        },
+        "type": "Footer",
+      },
+    ],
+  },
+  "type": "Container",
+}
+`);
+
+      // resolve the token balance, which triggers a re-render
+      resolveTokenBalancePromise();
+
+      // allow the event loop to run the re-render
+      await new Promise((resolve) => setTimeout(resolve, 0));
+
+      const confirmationContentWithBalance =
+        await confirmationShell.createConfirmationContent({
+          context: mockContext,
+          metadata: mockMetadata,
+          origin: mockOrigin,
+          chainId: 1,
+          scanDappUrlResult: null,
+          scanAddressResult: null,
+          existingPermissionsStatus: ExistingPermissionsState.None,
+          isGrantDisabled: false,
+        });
+
+      // concrete token balance, skeleton for fiat balance
+      expect(confirmationContentWithBalance).toMatchInlineSnapshot(`
+{
+  "key": null,
+  "props": {
+    "children": [
+      {
+        "key": null,
+        "props": {
+          "children": {
+            "key": null,
+            "props": {
+              "children": [
+                {
+                  "key": null,
+                  "props": {
+                    "center": true,
+                    "children": [
+                      {
+                        "key": null,
+                        "props": {
+                          "children": "Permission request",
+                          "size": "lg",
+                        },
+                        "type": "Heading",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "children": "This site wants permissions to spend your tokens.",
+                        },
+                        "type": "Text",
+                      },
+                    ],
+                  },
+                  "type": "Box",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": {
+                      "key": null,
+                      "props": {
+                        "children": [
+                          {
+                            "key": null,
+                            "props": {
+                              "alignment": "space-between",
+                              "children": {
+                                "key": null,
+                                "props": {
+                                  "children": [
+                                    {
+                                      "key": null,
+                                      "props": {
+                                        "children": "Account",
+                                      },
+                                      "type": "Text",
+                                    },
+                                    {
+                                      "key": null,
+                                      "props": {
+                                        "children": {
+                                          "key": null,
+                                          "props": {
+                                            "color": "muted",
+                                            "name": "question",
+                                            "size": "inherit",
+                                          },
+                                          "type": "Icon",
+                                        },
+                                        "content": {
+                                          "key": null,
+                                          "props": {
+                                            "children": "The account from which the permission is being granted.",
+                                          },
+                                          "type": "Text",
+                                        },
+                                      },
+                                      "type": "Tooltip",
+                                    },
+                                  ],
+                                  "direction": "horizontal",
+                                },
+                                "type": "Box",
+                              },
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "chainIds": [
+                                "eip155:1",
+                              ],
+                              "name": "account-selector",
+                              "switchGlobalAccount": false,
+                              "value": "eip155:1:0x1234567890123456789012345678901234567890",
+                            },
+                            "type": "AccountSelector",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "children": "This account will be upgraded to a smart account to complete this permission.",
+                              "color": "warning",
+                              "size": "sm",
+                            },
+                            "type": "Text",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "alignment": "end",
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {},
+                                  "type": "Skeleton",
+                                },
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": [
+                                      "1",
+                                      " ",
+                                      "available",
+                                    ],
+                                  },
+                                  "type": "Text",
+                                },
+                              ],
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                        ],
+                        "direction": "vertical",
+                      },
+                      "type": "Box",
+                    },
+                  },
+                  "type": "Section",
+                },
+                false,
+                false,
+                {
+                  "key": null,
+                  "props": {
+                    "children": {
+                      "key": null,
+                      "props": {
+                        "alignment": "space-between",
+                        "children": [
+                          {
+                            "key": null,
+                            "props": {
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "Justification",
+                                  },
+                                  "type": "Text",
+                                },
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": {
+                                      "key": null,
+                                      "props": {
+                                        "color": "muted",
+                                        "name": "question",
+                                        "size": "inherit",
+                                      },
+                                      "type": "Icon",
+                                    },
+                                    "content": {
+                                      "key": null,
+                                      "props": {
+                                        "children": "Justification given by the recipient for requesting this permission.",
+                                      },
+                                      "type": "Text",
+                                    },
+                                  },
+                                  "type": "Tooltip",
+                                },
+                              ],
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "Test justification text that is longer than twenty characters",
+                                  },
+                                  "type": "Text",
+                                },
+                                null,
+                              ],
+                              "direction": "vertical",
+                            },
+                            "type": "Box",
+                          },
+                        ],
+                        "direction": "vertical",
+                      },
+                      "type": "Box",
+                    },
+                  },
+                  "type": "Section",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": [
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Request from",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The site requesting the permission",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "alignment": "end",
+                                      "children": "https://example.com",
+                                    },
+                                    "type": "Text",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Recipient",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The site requesting the permission",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": {
+                                        "key": null,
+                                        "props": {
+                                          "children": "0x12345...67890",
+                                        },
+                                        "type": "Text",
+                                      },
+                                      "content": "0x1234567890123456789012345678901234567890",
+                                    },
+                                    "type": "Tooltip",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Network",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The network on which the permission is being requested",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "alignment": "end",
+                                      "children": "Ethereum Mainnet",
+                                    },
+                                    "type": "Text",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Token",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The token being requested",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": {
+                                        "key": null,
+                                        "props": {
+                                          "children": "ETH",
+                                          "href": "https://etherscan.io/address/0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3",
+                                        },
+                                        "type": "Link",
+                                      },
+                                      "content": "0x38c4A...611F3",
+                                    },
+                                    "type": "Tooltip",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      null,
+                      null,
+                    ],
+                  },
+                  "type": "Section",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": "Permission body",
+                  },
+                  "type": "Text",
+                },
+              ],
+              "direction": "vertical",
+            },
+            "type": "Box",
+          },
+        },
+        "type": "Box",
+      },
+      {
+        "key": null,
+        "props": {
+          "children": [
+            {
+              "key": null,
+              "props": {
+                "children": "Cancel",
+                "name": "cancel-button",
+                "variant": "destructive",
+              },
+              "type": "Button",
+            },
+            {
+              "key": null,
+              "props": {
+                "children": "Grant",
+                "disabled": false,
+                "name": "grant-button",
+                "variant": "primary",
+              },
+              "type": "Button",
+            },
+          ],
+        },
+        "type": "Footer",
+      },
+    ],
+  },
+  "type": "Container",
+}
+`);
+
+      // resolve the fiat balance, which triggers a re-render
+      resolveFiatBalancePromise();
+
+      // allow the event loop to run the re-render
+      await new Promise((resolve) => setTimeout(resolve, 0));
+
+      const confirmationContentWithFiatBalance =
+        await confirmationShell.createConfirmationContent({
+          context: mockContext,
+          metadata: mockMetadata,
+          origin: mockOrigin,
+          chainId: 1,
+          scanDappUrlResult: null,
+          scanAddressResult: null,
+          existingPermissionsStatus: ExistingPermissionsState.None,
+          isGrantDisabled: false,
+        });
+
+      // concrete token balance, concrete fiat balance
+      expect(confirmationContentWithFiatBalance).toMatchInlineSnapshot(`
+{
+  "key": null,
+  "props": {
+    "children": [
+      {
+        "key": null,
+        "props": {
+          "children": {
+            "key": null,
+            "props": {
+              "children": [
+                {
+                  "key": null,
+                  "props": {
+                    "center": true,
+                    "children": [
+                      {
+                        "key": null,
+                        "props": {
+                          "children": "Permission request",
+                          "size": "lg",
+                        },
+                        "type": "Heading",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "children": "This site wants permissions to spend your tokens.",
+                        },
+                        "type": "Text",
+                      },
+                    ],
+                  },
+                  "type": "Box",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": {
+                      "key": null,
+                      "props": {
+                        "children": [
+                          {
+                            "key": null,
+                            "props": {
+                              "alignment": "space-between",
+                              "children": {
+                                "key": null,
+                                "props": {
+                                  "children": [
+                                    {
+                                      "key": null,
+                                      "props": {
+                                        "children": "Account",
+                                      },
+                                      "type": "Text",
+                                    },
+                                    {
+                                      "key": null,
+                                      "props": {
+                                        "children": {
+                                          "key": null,
+                                          "props": {
+                                            "color": "muted",
+                                            "name": "question",
+                                            "size": "inherit",
+                                          },
+                                          "type": "Icon",
+                                        },
+                                        "content": {
+                                          "key": null,
+                                          "props": {
+                                            "children": "The account from which the permission is being granted.",
+                                          },
+                                          "type": "Text",
+                                        },
+                                      },
+                                      "type": "Tooltip",
+                                    },
+                                  ],
+                                  "direction": "horizontal",
+                                },
+                                "type": "Box",
+                              },
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "chainIds": [
+                                "eip155:1",
+                              ],
+                              "name": "account-selector",
+                              "switchGlobalAccount": false,
+                              "value": "eip155:1:0x1234567890123456789012345678901234567890",
+                            },
+                            "type": "AccountSelector",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "children": "This account will be upgraded to a smart account to complete this permission.",
+                              "color": "warning",
+                              "size": "sm",
+                            },
+                            "type": "Text",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "alignment": "end",
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "$1000",
+                                  },
+                                  "type": "Text",
+                                },
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": [
+                                      "1",
+                                      " ",
+                                      "available",
+                                    ],
+                                  },
+                                  "type": "Text",
+                                },
+                              ],
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                        ],
+                        "direction": "vertical",
+                      },
+                      "type": "Box",
+                    },
+                  },
+                  "type": "Section",
+                },
+                false,
+                false,
+                {
+                  "key": null,
+                  "props": {
+                    "children": {
+                      "key": null,
+                      "props": {
+                        "alignment": "space-between",
+                        "children": [
+                          {
+                            "key": null,
+                            "props": {
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "Justification",
+                                  },
+                                  "type": "Text",
+                                },
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": {
+                                      "key": null,
+                                      "props": {
+                                        "color": "muted",
+                                        "name": "question",
+                                        "size": "inherit",
+                                      },
+                                      "type": "Icon",
+                                    },
+                                    "content": {
+                                      "key": null,
+                                      "props": {
+                                        "children": "Justification given by the recipient for requesting this permission.",
+                                      },
+                                      "type": "Text",
+                                    },
+                                  },
+                                  "type": "Tooltip",
+                                },
+                              ],
+                              "direction": "horizontal",
+                            },
+                            "type": "Box",
+                          },
+                          {
+                            "key": null,
+                            "props": {
+                              "children": [
+                                {
+                                  "key": null,
+                                  "props": {
+                                    "children": "Test justification text that is longer than twenty characters",
+                                  },
+                                  "type": "Text",
+                                },
+                                null,
+                              ],
+                              "direction": "vertical",
+                            },
+                            "type": "Box",
+                          },
+                        ],
+                        "direction": "vertical",
+                      },
+                      "type": "Box",
+                    },
+                  },
+                  "type": "Section",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": [
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Request from",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The site requesting the permission",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "alignment": "end",
+                                      "children": "https://example.com",
+                                    },
+                                    "type": "Text",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Recipient",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The site requesting the permission",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": {
+                                        "key": null,
+                                        "props": {
+                                          "children": "0x12345...67890",
+                                        },
+                                        "type": "Text",
+                                      },
+                                      "content": "0x1234567890123456789012345678901234567890",
+                                    },
+                                    "type": "Tooltip",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Network",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The network on which the permission is being requested",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "alignment": "end",
+                                      "children": "Ethereum Mainnet",
+                                    },
+                                    "type": "Text",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      {
+                        "key": null,
+                        "props": {
+                          "alignment": "space-between",
+                          "children": [
+                            {
+                              "key": null,
+                              "props": {
+                                "alignment": "space-between",
+                                "children": [
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": [
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": "Token",
+                                          },
+                                          "type": "Text",
+                                        },
+                                        {
+                                          "key": null,
+                                          "props": {
+                                            "children": {
+                                              "key": null,
+                                              "props": {
+                                                "color": "muted",
+                                                "name": "question",
+                                                "size": "inherit",
+                                              },
+                                              "type": "Icon",
+                                            },
+                                            "content": {
+                                              "key": null,
+                                              "props": {
+                                                "children": "The token being requested",
+                                              },
+                                              "type": "Text",
+                                            },
+                                          },
+                                          "type": "Tooltip",
+                                        },
+                                      ],
+                                      "direction": "horizontal",
+                                    },
+                                    "type": "Box",
+                                  },
+                                  null,
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                            {
+                              "key": null,
+                              "props": {
+                                "children": [
+                                  null,
+                                  {
+                                    "key": null,
+                                    "props": {
+                                      "children": {
+                                        "key": null,
+                                        "props": {
+                                          "children": "ETH",
+                                          "href": "https://etherscan.io/address/0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3",
+                                        },
+                                        "type": "Link",
+                                      },
+                                      "content": "0x38c4A...611F3",
+                                    },
+                                    "type": "Tooltip",
+                                  },
+                                ],
+                                "direction": "horizontal",
+                              },
+                              "type": "Box",
+                            },
+                          ],
+                          "direction": "horizontal",
+                        },
+                        "type": "Box",
+                      },
+                      null,
+                      null,
+                    ],
+                  },
+                  "type": "Section",
+                },
+                {
+                  "key": null,
+                  "props": {
+                    "children": "Permission body",
+                  },
+                  "type": "Text",
+                },
+              ],
+              "direction": "vertical",
+            },
+            "type": "Box",
+          },
+        },
+        "type": "Box",
+      },
+      {
+        "key": null,
+        "props": {
+          "children": [
+            {
+              "key": null,
+              "props": {
+                "children": "Cancel",
+                "name": "cancel-button",
+                "variant": "destructive",
+              },
+              "type": "Button",
+            },
+            {
+              "key": null,
+              "props": {
+                "children": "Grant",
+                "disabled": false,
+                "name": "grant-button",
+                "variant": "primary",
+              },
+              "type": "Button",
+            },
+          ],
+        },
+        "type": "Footer",
+      },
+    ],
+  },
+  "type": "Container",
+}
+`);
+    });
+
+    it('cancels the balance loading when the account is changed', async () => {
+      const {
+        confirmationShell,
+        rules,
+        getBoundEvent,
+        tokenMetadataService,
+        updateContext,
+      } = setupTest();
+
+      let resolveTokenBalancePromise: () => void = (): void => {
+        throw new Error('Function should never be called');
+      };
+      const tokenBalancePromise = new Promise<TokenBalanceAndMetadata>(
+        (resolve) => {
+          resolveTokenBalancePromise = (): void =>
+            resolve(mockTokenBalanceAndMetadata);
+        },
+      );
+
+      tokenMetadataService.getTokenBalanceAndMetadata.mockReturnValueOnce(
+        tokenBalancePromise,
+      );
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: mockContext,
+        rules,
+        updateContext,
+      });
+
+      const accountSelectorChangeHandler = getBoundEvent({
+        elementName: 'account-selector',
+        eventType: 'InputChangeEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(accountSelectorChangeHandler).toBeDefined();
+
+      const mockAddress2Caip10 = `eip155:1:${mockAddress2}`;
+
+      await accountSelectorChangeHandler?.({
+        event: {
+          value: { addresses: [mockAddress2Caip10] } as any,
+          name: 'account-selector',
+          type: UserInputEventType.InputChangeEvent,
+        },
+        interfaceId: mockInterfaceId,
+      });
+
+      resolveTokenBalancePromise();
+      await new Promise((resolve) => setTimeout(resolve, 0));
+
+      // update context is called 4 times:
+      // 1. After the account is changed
+      // 2. After the account upgrade status is fetched for the new account
+      // 3. After the token balance is resolved for the second account
+      // 4. After the fiat balance is resolved for the second account
+      // if we didn't cancel the original balance loading, there would be another 2 instances
+      // 5. After the original token balance is resolved for the first account
+      // 6. After the original fiat balance is resolved for the first account
+      expect(updateContext).toHaveBeenCalledTimes(4);
+    });
+
+    it('unbinds the event handlers when the confirmation is resolved', async () => {
+      const {
+        confirmationShell,
+        rules,
+        getUnboundEvent,
+        getBoundEvent,
+        updateContext,
+      } = setupTest();
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: mockContext,
+        rules,
+        updateContext,
+      });
+
+      const accountSelectorBoundEvent = getBoundEvent({
+        elementName: 'account-selector',
+        eventType: 'InputChangeEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      const showMoreButtonBoundEvent = getBoundEvent({
+        elementName: 'show-more-justification',
+        eventType: 'ButtonClickEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(accountSelectorBoundEvent).toBeDefined();
+      expect(showMoreButtonBoundEvent).toBeDefined();
+      confirmationShell.resolveSession();
+
+      const accountSelectorUnboundEvent = getUnboundEvent({
+        elementName: 'account-selector',
+        eventType: 'InputChangeEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      const showMoreButtonUnboundEvent = getUnboundEvent({
+        elementName: 'show-more-justification',
+        eventType: 'ButtonClickEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(accountSelectorUnboundEvent).toBeDefined();
+      expect(showMoreButtonUnboundEvent).toBeDefined();
+    });
+
+    it('binds rule handlers even when isAdjustmentAllowed is false', async () => {
+      const rule: RuleDefinition<TestContextType, TestMetadataType> = {
+        name: 'amountPerSecond',
+        label: 'amountLabel',
+        type: 'number',
+        getRuleData: () => ({
+          value: '0x1',
+          isVisible: true,
+          isEditable: false,
+        }),
+        updateContext: (context) => context,
+      };
+      const { confirmationShell, rules, getBoundEvent, updateContext } =
+        setupTest({ rules: [rule] });
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: mockContext,
+        rules,
+        updateContext,
+      });
+
+      // Try to get a rule input handler - it should still be bound
+      const ruleInputHandler = getBoundEvent({
+        elementName: 'amountPerSecond', // Example rule input field
+        eventType: 'InputChangeEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(ruleInputHandler).toBeDefined();
+
+      // Account selector should still be bound (it's allowed even when adjustment is not allowed)
+      const accountSelectorHandler = getBoundEvent({
+        elementName: 'account-selector',
+        eventType: 'InputChangeEvent',
+        interfaceId: mockInterfaceId,
+      });
+
+      expect(accountSelectorHandler).toBeDefined();
+    });
+
+    it('skips balance fetch when context has no asset address', () => {
+      const { confirmationShell, updateContext, tokenMetadataService, rules } =
+        setupTest();
+
+      confirmationShell.bindSessionEvents({
+        interfaceId: mockInterfaceId,
+        initialContext: {
+          ...mockContext,
+          tokenAddressCaip19: NO_ASSET_ADDRESS,
+        },
+        rules,
+        updateContext,
+      });
+
+      expect(
+        tokenMetadataService.getTokenBalanceAndMetadata,
+      ).not.toHaveBeenCalled();
+    });
+  });
+});
