+++ packages/gator-permissions-snap/test/core/permissionHandler.test.ts
@@ -2,11 +2,9 @@ import { describe, expect, it, jest } from '@jest/globals';
 import type { PermissionRequest } from '@metamask/7715-permissions-shared/types';
 import { UserInputEventType } from '@metamask/snaps-sdk';
 
-import { AddressScanResultType } from '../../src/clients/trustSignalsClient';
 import type { TokenBalanceAndMetadata } from '../../src/clients/types';
 import type { AccountController } from '../../src/core/accountController';
 import { ExistingPermissionsState } from '../../src/core/existingpermissions/existingPermissionsState';
-import { METAMASK_FACILITATOR_ADDRESSES } from '../../src/core/facilitatorAddresses';
 import { PermissionHandler } from '../../src/core/permissionHandler';
 import type { PermissionRequestLifecycleOrchestrator } from '../../src/core/permissionRequestLifecycleOrchestrator';
 import type {
@@ -27,7 +25,6 @@ import type { MessageKey } from '../../src/utils/i18n';
 const mockAddress = '0x1234567890123456789012345678901234567890' as const;
 const mockAddress2 = '0x1234567890123456789012345678901234567891' as const;
 const mockAssetAddress = '0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3' as const;
-const mockInterfaceId = 'test-interface-id';
 const mockOrigin = 'https://example.com';
 const mockTokenBalanceFiat = '$1000';
 
@@ -400,3637 +397,28 @@ describe('PermissionHandler', () => {
   });
 
   describe('lifecycleHandlers', () => {
-    describe('createConfirmationContent', () => {
-      it('calls createConfirmationContent to produce the permission specific content', async () => {
-        const { permissionHandler, getLifecycleHandlers, dependencies } =
-          setupTest();
+    it('delegates confirmation shell methods through lifecycle handlers', async () => {
+      const { permissionHandler, getLifecycleHandlers, dependencies } =
+        setupTest();
 
-        await permissionHandler.handlePermissionRequest(mockOrigin);
+      await permissionHandler.handlePermissionRequest(mockOrigin);
 
-        const lifecycleHandlers = getLifecycleHandlers();
+      const lifecycleHandlers = getLifecycleHandlers();
 
-        await lifecycleHandlers.createConfirmationContent({
-          context: mockContext,
-          metadata: mockMetadata,
-          origin: mockOrigin,
-          chainId: 1,
-          scanDappUrlResult: null,
-          scanAddressResult: null,
-          existingPermissionsStatus: ExistingPermissionsState.None,
-          isGrantDisabled: false,
-        });
-
-        expect(dependencies.createConfirmationContent).toHaveBeenCalledWith({
-          context: mockContext,
-          metadata: mockMetadata,
-        });
+      await lifecycleHandlers.createConfirmationContent({
+        context: mockContext,
+        metadata: mockMetadata,
+        origin: mockOrigin,
+        chainId: 1,
+        scanDappUrlResult: null,
+        scanAddressResult: null,
+        existingPermissionsStatus: ExistingPermissionsState.None,
+        isGrantDisabled: false,
       });
 
-      it('creates a PermissionHandlerContent', async () => {
-        const { permissionHandler, getLifecycleHandlers } = setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        const result = await lifecycleHandlers.createConfirmationContent({
-          context: mockContext,
-          metadata: mockMetadata,
-          origin: mockOrigin,
-          chainId: 1,
-          scanDappUrlResult: null,
-          scanAddressResult: null,
-          existingPermissionsStatus: ExistingPermissionsState.None,
-          isGrantDisabled: false,
-        });
-
-        expect(result).toBeDefined();
-        expect(result.type).toBe('Container');
-      });
-
-      it('uses translated fallback for address warning when scanAddressResult.label is empty', async () => {
-        const { permissionHandler, getLifecycleHandlers } = setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        const result = await lifecycleHandlers.createConfirmationContent({
-          context: mockContext,
-          metadata: mockMetadata,
-          origin: mockOrigin,
-          chainId: 1,
-          scanDappUrlResult: null,
-          scanAddressResult: {
-            resultType: AddressScanResultType.Malicious,
-            label: '',
-          },
-          existingPermissionsStatus: ExistingPermissionsState.None,
-          isGrantDisabled: false,
-        });
-
-        // When label is empty, permissionHandlerContent should use t('maliciousAddressLabel')
-        const serialized = JSON.stringify(result);
-        expect(serialized).toContain('Malicious address');
-      });
-
-      it('renders redeemer addresses as redeemers when no payee rule is present', async () => {
-        const { permissionHandler, getLifecycleHandlers } = setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        const result = await lifecycleHandlers.createConfirmationContent({
-          context: {
-            ...mockContext,
-            redeemerAddresses: ['0x1111111111111111111111111111111111111111'],
-          },
-          metadata: mockMetadata,
-          origin: mockOrigin,
-          chainId: 1,
-          scanDappUrlResult: null,
-          scanAddressResult: null,
-          existingPermissionsStatus: ExistingPermissionsState.None,
-          isGrantDisabled: false,
-        });
-
-        const serialized = JSON.stringify(result);
-        expect(serialized).toContain('Redeemers');
-        expect(serialized).not.toContain('Facilitators');
-      });
-
-      it('renders known MetaMask facilitator redeemer addresses as a MetaMask facilitator redeemer', async () => {
-        const { permissionHandler, getLifecycleHandlers } = setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        const result = await lifecycleHandlers.createConfirmationContent({
-          context: {
-            ...mockContext,
-            redeemerAddresses: [...METAMASK_FACILITATOR_ADDRESSES],
-          },
-          metadata: mockMetadata,
-          origin: mockOrigin,
-          chainId: 1,
-          scanDappUrlResult: null,
-          scanAddressResult: null,
-          existingPermissionsStatus: ExistingPermissionsState.None,
-          isGrantDisabled: false,
-        });
-
-        const serialized = JSON.stringify(result);
-        expect(serialized).toContain('Redeemers');
-        expect(serialized).toContain('MetaMask facilitator');
-        expect(serialized).toContain(
-          'Only these addresses may redeem this permission.',
-        );
-        expect(serialized).not.toContain('Facilitators');
-      });
-
-      it('does not render arbitrary redeemer addresses as facilitators when a payee rule is present', async () => {
-        const { permissionHandler, getLifecycleHandlers } = setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        const result = await lifecycleHandlers.createConfirmationContent({
-          context: {
-            ...mockContext,
-            redeemerAddresses: ['0x1111111111111111111111111111111111111111'],
-            payeeAddresses: ['0x2222222222222222222222222222222222222222'],
-          },
-          metadata: mockMetadata,
-          origin: mockOrigin,
-          chainId: 1,
-          scanDappUrlResult: null,
-          scanAddressResult: null,
-          existingPermissionsStatus: ExistingPermissionsState.None,
-          isGrantDisabled: false,
-        });
-
-        const serialized = JSON.stringify(result);
-        expect(serialized).toContain('Redeemers');
-        expect(serialized).not.toContain('Facilitators');
-      });
-    });
-
-    describe('onConfirmationCreated', () => {
-      it('registers event handlers for account selection and justification toggle', async () => {
-        const rule: RuleDefinition<TestContextType, TestMetadataType> = {
-          name: 'amountPerSecond',
-          label: 'amountLabel',
-          type: 'number',
-          getRuleData: () => ({
-            value: '0x1',
-            isVisible: true,
-            isEditable: false,
-          }),
-          updateContext: (context) => context,
-        };
-        const {
-          permissionHandler,
-          getLifecycleHandlers,
-          getBoundEvent,
-          updateContext,
-        } = setupTest({ rules: [rule] });
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        lifecycleHandlers.onConfirmationCreated?.({
-          interfaceId: mockInterfaceId,
-          initialContext: mockContext,
-          updateContext,
-        });
-
-        const accountSelectorBoundEvent = getBoundEvent({
-          elementName: 'account-selector',
-          eventType: 'InputChangeEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        const showMoreButtonBoundEvent = getBoundEvent({
-          elementName: 'show-more-justification',
-          eventType: 'ButtonClickEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(accountSelectorBoundEvent).toBeDefined();
-        expect(showMoreButtonBoundEvent).toBeDefined();
-      });
-
-      it('loads the balance for the selected account', async () => {
-        const {
-          permissionHandler,
-          getLifecycleHandlers,
-          tokenMetadataService,
-          updateContext,
-        } = setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        lifecycleHandlers.onConfirmationCreated?.({
-          interfaceId: mockInterfaceId,
-          initialContext: mockContext,
-          updateContext,
-        });
-
-        expect(
-          tokenMetadataService.getTokenBalanceAndMetadata,
-        ).toHaveBeenCalledWith({
-          chainId: 1,
-          account: mockAddress,
-          assetAddress: mockAssetAddress,
-        });
-      });
-
-      it('updates the context when the account is changed', async () => {
-        const {
-          permissionHandler,
-          getLifecycleHandlers,
-          getBoundEvent,
-          updateContext,
-        } = setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        lifecycleHandlers.onConfirmationCreated?.({
-          interfaceId: mockInterfaceId,
-          initialContext: mockContext,
-          updateContext,
-        });
-
-        const accountSelectorChangeHandler = getBoundEvent({
-          elementName: 'account-selector',
-          eventType: 'InputChangeEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(accountSelectorChangeHandler).toBeDefined();
-
-        const mockAddress2Caip10 = `eip155:1:${mockAddress2}`;
-
-        await accountSelectorChangeHandler?.({
-          event: {
-            value: { addresses: [mockAddress2Caip10] } as any,
-            name: 'account-selector',
-            type: UserInputEventType.InputChangeEvent,
-          },
-          interfaceId: mockInterfaceId,
-        });
-
-        const expectedUpdatedContext = {
-          ...mockContext,
-          accountAddressCaip10: mockAddress2Caip10,
-        };
-
-        expect(updateContext).toHaveBeenCalledWith({
-          updatedContext: expectedUpdatedContext,
-        });
-      });
-
-      it('updates the balance when the account is changed', async () => {
-        const {
-          permissionHandler,
-          getLifecycleHandlers,
-          getBoundEvent,
-          tokenMetadataService,
-          updateContext,
-        } = setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        lifecycleHandlers.onConfirmationCreated?.({
-          interfaceId: mockInterfaceId,
-          initialContext: mockContext,
-          updateContext,
-        });
-
-        const accountSelectorChangeHandler = getBoundEvent({
-          elementName: 'account-selector',
-          eventType: 'InputChangeEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(accountSelectorChangeHandler).toBeDefined();
-
-        const mockAddress2Caip10 = `eip155:1:${mockAddress2}`;
-
-        await accountSelectorChangeHandler?.({
-          event: {
-            value: { addresses: [mockAddress2Caip10] } as any,
-            name: 'account-selector',
-            type: UserInputEventType.InputChangeEvent,
-          },
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(
-          tokenMetadataService.getTokenBalanceAndMetadata,
-        ).toHaveBeenCalledTimes(2);
-
-        expect(
-          tokenMetadataService.getTokenBalanceAndMetadata,
-        ).toHaveBeenCalledWith({
-          chainId: 1,
-          account: mockAddress2,
-          assetAddress: mockAssetAddress,
-        });
-      });
-
-      it('renders the balance in the confirmation content', async () => {
-        const { permissionHandler, getLifecycleHandlers, updateContext } =
-          setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        lifecycleHandlers.onConfirmationCreated?.({
-          interfaceId: mockInterfaceId,
-          initialContext: mockContext,
-          updateContext,
-        });
-
-        const confirmationContent =
-          await lifecycleHandlers.createConfirmationContent({
-            context: mockContext,
-            metadata: mockMetadata,
-            origin: mockOrigin,
-            chainId: 1,
-            scanDappUrlResult: null,
-            scanAddressResult: null,
-            existingPermissionsStatus: ExistingPermissionsState.None,
-            isGrantDisabled: false,
-          });
-        expect(confirmationContent).toMatchInlineSnapshot(`
-{
-  "key": null,
-  "props": {
-    "children": [
-      {
-        "key": null,
-        "props": {
-          "children": {
-            "key": null,
-            "props": {
-              "children": [
-                {
-                  "key": null,
-                  "props": {
-                    "center": true,
-                    "children": [
-                      {
-                        "key": null,
-                        "props": {
-                          "children": "Permission request",
-                          "size": "lg",
-                        },
-                        "type": "Heading",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "children": "This site wants permissions to spend your tokens.",
-                        },
-                        "type": "Text",
-                      },
-                    ],
-                  },
-                  "type": "Box",
-                },
-                {
-                  "key": null,
-                  "props": {
-                    "children": {
-                      "key": null,
-                      "props": {
-                        "children": [
-                          {
-                            "key": null,
-                            "props": {
-                              "alignment": "space-between",
-                              "children": {
-                                "key": null,
-                                "props": {
-                                  "children": [
-                                    {
-                                      "key": null,
-                                      "props": {
-                                        "children": "Account",
-                                      },
-                                      "type": "Text",
-                                    },
-                                    {
-                                      "key": null,
-                                      "props": {
-                                        "children": {
-                                          "key": null,
-                                          "props": {
-                                            "color": "muted",
-                                            "name": "question",
-                                            "size": "inherit",
-                                          },
-                                          "type": "Icon",
-                                        },
-                                        "content": {
-                                          "key": null,
-                                          "props": {
-                                            "children": "The account from which the permission is being granted.",
-                                          },
-                                          "type": "Text",
-                                        },
-                                      },
-                                      "type": "Tooltip",
-                                    },
-                                  ],
-                                  "direction": "horizontal",
-                                },
-                                "type": "Box",
-                              },
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "chainIds": [
-                                "eip155:1",
-                              ],
-                              "name": "account-selector",
-                              "switchGlobalAccount": false,
-                              "value": "eip155:1:0x1234567890123456789012345678901234567890",
-                            },
-                            "type": "AccountSelector",
-                          },
-                          false,
-                          {
-                            "key": null,
-                            "props": {
-                              "alignment": "end",
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {},
-                                  "type": "Skeleton",
-                                },
-                                {
-                                  "key": null,
-                                  "props": {},
-                                  "type": "Skeleton",
-                                },
-                              ],
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                        ],
-                        "direction": "vertical",
-                      },
-                      "type": "Box",
-                    },
-                  },
-                  "type": "Section",
-                },
-                false,
-                false,
-                {
-                  "key": null,
-                  "props": {
-                    "children": {
-                      "key": null,
-                      "props": {
-                        "alignment": "space-between",
-                        "children": [
-                          {
-                            "key": null,
-                            "props": {
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "Justification",
-                                  },
-                                  "type": "Text",
-                                },
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": {
-                                      "key": null,
-                                      "props": {
-                                        "color": "muted",
-                                        "name": "question",
-                                        "size": "inherit",
-                                      },
-                                      "type": "Icon",
-                                    },
-                                    "content": {
-                                      "key": null,
-                                      "props": {
-                                        "children": "Justification given by the recipient for requesting this permission.",
-                                      },
-                                      "type": "Text",
-                                    },
-                                  },
-                                  "type": "Tooltip",
-                                },
-                              ],
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "Test justification text that is longer than twenty characters",
-                                  },
-                                  "type": "Text",
-                                },
-                                null,
-                              ],
-                              "direction": "vertical",
-                            },
-                            "type": "Box",
-                          },
-                        ],
-                        "direction": "vertical",
-                      },
-                      "type": "Box",
-                    },
-                  },
-                  "type": "Section",
-                },
-                {
-                  "key": null,
-                  "props": {
-                    "children": [
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Request from",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The site requesting the permission",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "alignment": "end",
-                                      "children": "https://example.com",
-                                    },
-                                    "type": "Text",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Recipient",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The site requesting the permission",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": {
-                                        "key": null,
-                                        "props": {
-                                          "children": "0x12345...67890",
-                                        },
-                                        "type": "Text",
-                                      },
-                                      "content": "0x1234567890123456789012345678901234567890",
-                                    },
-                                    "type": "Tooltip",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Network",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The network on which the permission is being requested",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "alignment": "end",
-                                      "children": "Ethereum Mainnet",
-                                    },
-                                    "type": "Text",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Token",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The token being requested",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": {
-                                        "key": null,
-                                        "props": {
-                                          "children": "ETH",
-                                          "href": "https://etherscan.io/address/0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3",
-                                        },
-                                        "type": "Link",
-                                      },
-                                      "content": "0x38c4A...611F3",
-                                    },
-                                    "type": "Tooltip",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      null,
-                      null,
-                    ],
-                  },
-                  "type": "Section",
-                },
-                undefined,
-              ],
-              "direction": "vertical",
-            },
-            "type": "Box",
-          },
-        },
-        "type": "Box",
-      },
-      {
-        "key": null,
-        "props": {
-          "children": [
-            {
-              "key": null,
-              "props": {
-                "children": "Cancel",
-                "name": "cancel-button",
-                "variant": "destructive",
-              },
-              "type": "Button",
-            },
-            {
-              "key": null,
-              "props": {
-                "children": "Grant",
-                "disabled": false,
-                "name": "grant-button",
-                "variant": "primary",
-              },
-              "type": "Button",
-            },
-          ],
-        },
-        "type": "Footer",
-      },
-    ],
-  },
-  "type": "Container",
-}
-`);
-      });
-
-      it('updates the balance in the confirmation content when the account is changed', async () => {
-        const {
-          permissionHandler,
-          getLifecycleHandlers,
-          getBoundEvent,
-          tokenPricesService,
-          updateContext,
-        } = setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        lifecycleHandlers.onConfirmationCreated?.({
-          interfaceId: mockInterfaceId,
-          initialContext: mockContext,
-          updateContext,
-        });
-
-        const accountSelectorChangeHandler = getBoundEvent({
-          elementName: 'account-selector',
-          eventType: 'InputChangeEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(accountSelectorChangeHandler).toBeDefined();
-
-        const mockAddress2Caip10 = `eip155:1:${mockAddress2}`;
-
-        tokenPricesService.getCryptoToFiatConversion.mockResolvedValue('$2000');
-
-        await accountSelectorChangeHandler?.({
-          event: {
-            value: { addresses: [mockAddress2Caip10] } as any,
-            name: 'account-selector',
-            type: UserInputEventType.InputChangeEvent,
-          },
-          interfaceId: mockInterfaceId,
-        });
-
-        const confirmationContent =
-          await lifecycleHandlers.createConfirmationContent({
-            context: mockContext,
-            metadata: mockMetadata,
-            origin: mockOrigin,
-            chainId: 1,
-            scanDappUrlResult: null,
-            scanAddressResult: null,
-            existingPermissionsStatus: ExistingPermissionsState.None,
-            isGrantDisabled: false,
-          });
-
-        expect(confirmationContent).toMatchInlineSnapshot(`
-{
-  "key": null,
-  "props": {
-    "children": [
-      {
-        "key": null,
-        "props": {
-          "children": {
-            "key": null,
-            "props": {
-              "children": [
-                {
-                  "key": null,
-                  "props": {
-                    "center": true,
-                    "children": [
-                      {
-                        "key": null,
-                        "props": {
-                          "children": "Permission request",
-                          "size": "lg",
-                        },
-                        "type": "Heading",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "children": "This site wants permissions to spend your tokens.",
-                        },
-                        "type": "Text",
-                      },
-                    ],
-                  },
-                  "type": "Box",
-                },
-                {
-                  "key": null,
-                  "props": {
-                    "children": {
-                      "key": null,
-                      "props": {
-                        "children": [
-                          {
-                            "key": null,
-                            "props": {
-                              "alignment": "space-between",
-                              "children": {
-                                "key": null,
-                                "props": {
-                                  "children": [
-                                    {
-                                      "key": null,
-                                      "props": {
-                                        "children": "Account",
-                                      },
-                                      "type": "Text",
-                                    },
-                                    {
-                                      "key": null,
-                                      "props": {
-                                        "children": {
-                                          "key": null,
-                                          "props": {
-                                            "color": "muted",
-                                            "name": "question",
-                                            "size": "inherit",
-                                          },
-                                          "type": "Icon",
-                                        },
-                                        "content": {
-                                          "key": null,
-                                          "props": {
-                                            "children": "The account from which the permission is being granted.",
-                                          },
-                                          "type": "Text",
-                                        },
-                                      },
-                                      "type": "Tooltip",
-                                    },
-                                  ],
-                                  "direction": "horizontal",
-                                },
-                                "type": "Box",
-                              },
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "chainIds": [
-                                "eip155:1",
-                              ],
-                              "name": "account-selector",
-                              "switchGlobalAccount": false,
-                              "value": "eip155:1:0x1234567890123456789012345678901234567890",
-                            },
-                            "type": "AccountSelector",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "children": "This account will be upgraded to a smart account to complete this permission.",
-                              "color": "warning",
-                              "size": "sm",
-                            },
-                            "type": "Text",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "alignment": "end",
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {},
-                                  "type": "Skeleton",
-                                },
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": [
-                                      "1",
-                                      " ",
-                                      "available",
-                                    ],
-                                  },
-                                  "type": "Text",
-                                },
-                              ],
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                        ],
-                        "direction": "vertical",
-                      },
-                      "type": "Box",
-                    },
-                  },
-                  "type": "Section",
-                },
-                false,
-                false,
-                {
-                  "key": null,
-                  "props": {
-                    "children": {
-                      "key": null,
-                      "props": {
-                        "alignment": "space-between",
-                        "children": [
-                          {
-                            "key": null,
-                            "props": {
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "Justification",
-                                  },
-                                  "type": "Text",
-                                },
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": {
-                                      "key": null,
-                                      "props": {
-                                        "color": "muted",
-                                        "name": "question",
-                                        "size": "inherit",
-                                      },
-                                      "type": "Icon",
-                                    },
-                                    "content": {
-                                      "key": null,
-                                      "props": {
-                                        "children": "Justification given by the recipient for requesting this permission.",
-                                      },
-                                      "type": "Text",
-                                    },
-                                  },
-                                  "type": "Tooltip",
-                                },
-                              ],
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "Test justification text that is longer than twenty characters",
-                                  },
-                                  "type": "Text",
-                                },
-                                null,
-                              ],
-                              "direction": "vertical",
-                            },
-                            "type": "Box",
-                          },
-                        ],
-                        "direction": "vertical",
-                      },
-                      "type": "Box",
-                    },
-                  },
-                  "type": "Section",
-                },
-                {
-                  "key": null,
-                  "props": {
-                    "children": [
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Request from",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The site requesting the permission",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "alignment": "end",
-                                      "children": "https://example.com",
-                                    },
-                                    "type": "Text",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Recipient",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The site requesting the permission",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": {
-                                        "key": null,
-                                        "props": {
-                                          "children": "0x12345...67890",
-                                        },
-                                        "type": "Text",
-                                      },
-                                      "content": "0x1234567890123456789012345678901234567890",
-                                    },
-                                    "type": "Tooltip",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Network",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The network on which the permission is being requested",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "alignment": "end",
-                                      "children": "Ethereum Mainnet",
-                                    },
-                                    "type": "Text",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Token",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The token being requested",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": {
-                                        "key": null,
-                                        "props": {
-                                          "children": "ETH",
-                                          "href": "https://etherscan.io/address/0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3",
-                                        },
-                                        "type": "Link",
-                                      },
-                                      "content": "0x38c4A...611F3",
-                                    },
-                                    "type": "Tooltip",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      null,
-                      null,
-                    ],
-                  },
-                  "type": "Section",
-                },
-                undefined,
-              ],
-              "direction": "vertical",
-            },
-            "type": "Box",
-          },
-        },
-        "type": "Box",
-      },
-      {
-        "key": null,
-        "props": {
-          "children": [
-            {
-              "key": null,
-              "props": {
-                "children": "Cancel",
-                "name": "cancel-button",
-                "variant": "destructive",
-              },
-              "type": "Button",
-            },
-            {
-              "key": null,
-              "props": {
-                "children": "Grant",
-                "disabled": false,
-                "name": "grant-button",
-                "variant": "primary",
-              },
-              "type": "Button",
-            },
-          ],
-        },
-        "type": "Footer",
-      },
-    ],
-  },
-  "type": "Container",
-}
-`);
-      });
-
-      it('renders skeletons while the balance is loading', async () => {
-        const {
-          permissionHandler,
-          getLifecycleHandlers,
-          getBoundEvent,
-          tokenMetadataService,
-          tokenPricesService,
-          updateContext,
-        } = setupTest();
-
-        let resolveTokenBalancePromise: () => void = (): void => {
-          throw new Error('Function should never be called');
-        };
-        const tokenBalancePromise = new Promise<TokenBalanceAndMetadata>(
-          (resolve) => {
-            resolveTokenBalancePromise = (): void =>
-              resolve(mockTokenBalanceAndMetadata);
-          },
-        );
-
-        tokenMetadataService.getTokenBalanceAndMetadata.mockReturnValue(
-          tokenBalancePromise,
-        );
-
-        let resolveFiatBalancePromise: () => void = (): void => {
-          throw new Error('Function should never be called');
-        };
-        const fiatBalancePromise = new Promise<string>((resolve) => {
-          resolveFiatBalancePromise = (): void => resolve(mockTokenBalanceFiat);
-        });
-        tokenPricesService.getCryptoToFiatConversion.mockReturnValue(
-          fiatBalancePromise,
-        );
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        lifecycleHandlers.onConfirmationCreated?.({
-          interfaceId: mockInterfaceId,
-          initialContext: mockContext,
-          updateContext,
-        });
-
-        const accountSelectorChangeHandler = getBoundEvent({
-          elementName: 'account-selector',
-          eventType: 'InputChangeEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(accountSelectorChangeHandler).toBeDefined();
-
-        const mockAddress2Caip10 = `eip155:1:${mockAddress2}`;
-
-        await accountSelectorChangeHandler?.({
-          event: {
-            value: { addresses: [mockAddress2Caip10] } as any,
-            name: 'account-selector',
-            type: UserInputEventType.InputChangeEvent,
-          },
-          interfaceId: mockInterfaceId,
-        });
-
-        // called twice:
-        // 1. with the context updated to include the new account address
-        // 2. after the account upgrade status is fetched for the new account
-        expect(updateContext).toHaveBeenCalledTimes(2);
-        expect(updateContext).toHaveBeenCalledWith({
-          updatedContext: {
-            ...mockContext,
-            accountAddressCaip10: mockAddress2Caip10,
-          },
-        });
-
-        const confirmationContent =
-          await lifecycleHandlers.createConfirmationContent({
-            context: mockContext,
-            metadata: mockMetadata,
-            origin: mockOrigin,
-            chainId: 1,
-            scanDappUrlResult: null,
-            scanAddressResult: null,
-            existingPermissionsStatus: ExistingPermissionsState.None,
-            isGrantDisabled: false,
-          });
-
-        // skeletons in place of both the token balance and fiat balance
-        expect(confirmationContent).toMatchInlineSnapshot(`
-{
-  "key": null,
-  "props": {
-    "children": [
-      {
-        "key": null,
-        "props": {
-          "children": {
-            "key": null,
-            "props": {
-              "children": [
-                {
-                  "key": null,
-                  "props": {
-                    "center": true,
-                    "children": [
-                      {
-                        "key": null,
-                        "props": {
-                          "children": "Permission request",
-                          "size": "lg",
-                        },
-                        "type": "Heading",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "children": "This site wants permissions to spend your tokens.",
-                        },
-                        "type": "Text",
-                      },
-                    ],
-                  },
-                  "type": "Box",
-                },
-                {
-                  "key": null,
-                  "props": {
-                    "children": {
-                      "key": null,
-                      "props": {
-                        "children": [
-                          {
-                            "key": null,
-                            "props": {
-                              "alignment": "space-between",
-                              "children": {
-                                "key": null,
-                                "props": {
-                                  "children": [
-                                    {
-                                      "key": null,
-                                      "props": {
-                                        "children": "Account",
-                                      },
-                                      "type": "Text",
-                                    },
-                                    {
-                                      "key": null,
-                                      "props": {
-                                        "children": {
-                                          "key": null,
-                                          "props": {
-                                            "color": "muted",
-                                            "name": "question",
-                                            "size": "inherit",
-                                          },
-                                          "type": "Icon",
-                                        },
-                                        "content": {
-                                          "key": null,
-                                          "props": {
-                                            "children": "The account from which the permission is being granted.",
-                                          },
-                                          "type": "Text",
-                                        },
-                                      },
-                                      "type": "Tooltip",
-                                    },
-                                  ],
-                                  "direction": "horizontal",
-                                },
-                                "type": "Box",
-                              },
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "chainIds": [
-                                "eip155:1",
-                              ],
-                              "name": "account-selector",
-                              "switchGlobalAccount": false,
-                              "value": "eip155:1:0x1234567890123456789012345678901234567890",
-                            },
-                            "type": "AccountSelector",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "children": "This account will be upgraded to a smart account to complete this permission.",
-                              "color": "warning",
-                              "size": "sm",
-                            },
-                            "type": "Text",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "alignment": "end",
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {},
-                                  "type": "Skeleton",
-                                },
-                                {
-                                  "key": null,
-                                  "props": {},
-                                  "type": "Skeleton",
-                                },
-                              ],
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                        ],
-                        "direction": "vertical",
-                      },
-                      "type": "Box",
-                    },
-                  },
-                  "type": "Section",
-                },
-                false,
-                false,
-                {
-                  "key": null,
-                  "props": {
-                    "children": {
-                      "key": null,
-                      "props": {
-                        "alignment": "space-between",
-                        "children": [
-                          {
-                            "key": null,
-                            "props": {
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "Justification",
-                                  },
-                                  "type": "Text",
-                                },
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": {
-                                      "key": null,
-                                      "props": {
-                                        "color": "muted",
-                                        "name": "question",
-                                        "size": "inherit",
-                                      },
-                                      "type": "Icon",
-                                    },
-                                    "content": {
-                                      "key": null,
-                                      "props": {
-                                        "children": "Justification given by the recipient for requesting this permission.",
-                                      },
-                                      "type": "Text",
-                                    },
-                                  },
-                                  "type": "Tooltip",
-                                },
-                              ],
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "Test justification text that is longer than twenty characters",
-                                  },
-                                  "type": "Text",
-                                },
-                                null,
-                              ],
-                              "direction": "vertical",
-                            },
-                            "type": "Box",
-                          },
-                        ],
-                        "direction": "vertical",
-                      },
-                      "type": "Box",
-                    },
-                  },
-                  "type": "Section",
-                },
-                {
-                  "key": null,
-                  "props": {
-                    "children": [
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Request from",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The site requesting the permission",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "alignment": "end",
-                                      "children": "https://example.com",
-                                    },
-                                    "type": "Text",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Recipient",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The site requesting the permission",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": {
-                                        "key": null,
-                                        "props": {
-                                          "children": "0x12345...67890",
-                                        },
-                                        "type": "Text",
-                                      },
-                                      "content": "0x1234567890123456789012345678901234567890",
-                                    },
-                                    "type": "Tooltip",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Network",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The network on which the permission is being requested",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "alignment": "end",
-                                      "children": "Ethereum Mainnet",
-                                    },
-                                    "type": "Text",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Token",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The token being requested",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": {
-                                        "key": null,
-                                        "props": {
-                                          "children": "ETH",
-                                          "href": "https://etherscan.io/address/0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3",
-                                        },
-                                        "type": "Link",
-                                      },
-                                      "content": "0x38c4A...611F3",
-                                    },
-                                    "type": "Tooltip",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      null,
-                      null,
-                    ],
-                  },
-                  "type": "Section",
-                },
-                undefined,
-              ],
-              "direction": "vertical",
-            },
-            "type": "Box",
-          },
-        },
-        "type": "Box",
-      },
-      {
-        "key": null,
-        "props": {
-          "children": [
-            {
-              "key": null,
-              "props": {
-                "children": "Cancel",
-                "name": "cancel-button",
-                "variant": "destructive",
-              },
-              "type": "Button",
-            },
-            {
-              "key": null,
-              "props": {
-                "children": "Grant",
-                "disabled": false,
-                "name": "grant-button",
-                "variant": "primary",
-              },
-              "type": "Button",
-            },
-          ],
-        },
-        "type": "Footer",
-      },
-    ],
-  },
-  "type": "Container",
-}
-`);
-
-        // resolve the token balance, which triggers a re-render
-        resolveTokenBalancePromise();
-
-        // allow the event loop to run the re-render
-        await new Promise((resolve) => setTimeout(resolve, 0));
-
-        const confirmationContentWithBalance =
-          await lifecycleHandlers.createConfirmationContent({
-            context: mockContext,
-            metadata: mockMetadata,
-            origin: mockOrigin,
-            chainId: 1,
-            scanDappUrlResult: null,
-            scanAddressResult: null,
-            existingPermissionsStatus: ExistingPermissionsState.None,
-            isGrantDisabled: false,
-          });
-
-        // concrete token balance, skeleton for fiat balance
-        expect(confirmationContentWithBalance).toMatchInlineSnapshot(`
-{
-  "key": null,
-  "props": {
-    "children": [
-      {
-        "key": null,
-        "props": {
-          "children": {
-            "key": null,
-            "props": {
-              "children": [
-                {
-                  "key": null,
-                  "props": {
-                    "center": true,
-                    "children": [
-                      {
-                        "key": null,
-                        "props": {
-                          "children": "Permission request",
-                          "size": "lg",
-                        },
-                        "type": "Heading",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "children": "This site wants permissions to spend your tokens.",
-                        },
-                        "type": "Text",
-                      },
-                    ],
-                  },
-                  "type": "Box",
-                },
-                {
-                  "key": null,
-                  "props": {
-                    "children": {
-                      "key": null,
-                      "props": {
-                        "children": [
-                          {
-                            "key": null,
-                            "props": {
-                              "alignment": "space-between",
-                              "children": {
-                                "key": null,
-                                "props": {
-                                  "children": [
-                                    {
-                                      "key": null,
-                                      "props": {
-                                        "children": "Account",
-                                      },
-                                      "type": "Text",
-                                    },
-                                    {
-                                      "key": null,
-                                      "props": {
-                                        "children": {
-                                          "key": null,
-                                          "props": {
-                                            "color": "muted",
-                                            "name": "question",
-                                            "size": "inherit",
-                                          },
-                                          "type": "Icon",
-                                        },
-                                        "content": {
-                                          "key": null,
-                                          "props": {
-                                            "children": "The account from which the permission is being granted.",
-                                          },
-                                          "type": "Text",
-                                        },
-                                      },
-                                      "type": "Tooltip",
-                                    },
-                                  ],
-                                  "direction": "horizontal",
-                                },
-                                "type": "Box",
-                              },
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "chainIds": [
-                                "eip155:1",
-                              ],
-                              "name": "account-selector",
-                              "switchGlobalAccount": false,
-                              "value": "eip155:1:0x1234567890123456789012345678901234567890",
-                            },
-                            "type": "AccountSelector",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "children": "This account will be upgraded to a smart account to complete this permission.",
-                              "color": "warning",
-                              "size": "sm",
-                            },
-                            "type": "Text",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "alignment": "end",
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {},
-                                  "type": "Skeleton",
-                                },
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": [
-                                      "1",
-                                      " ",
-                                      "available",
-                                    ],
-                                  },
-                                  "type": "Text",
-                                },
-                              ],
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                        ],
-                        "direction": "vertical",
-                      },
-                      "type": "Box",
-                    },
-                  },
-                  "type": "Section",
-                },
-                false,
-                false,
-                {
-                  "key": null,
-                  "props": {
-                    "children": {
-                      "key": null,
-                      "props": {
-                        "alignment": "space-between",
-                        "children": [
-                          {
-                            "key": null,
-                            "props": {
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "Justification",
-                                  },
-                                  "type": "Text",
-                                },
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": {
-                                      "key": null,
-                                      "props": {
-                                        "color": "muted",
-                                        "name": "question",
-                                        "size": "inherit",
-                                      },
-                                      "type": "Icon",
-                                    },
-                                    "content": {
-                                      "key": null,
-                                      "props": {
-                                        "children": "Justification given by the recipient for requesting this permission.",
-                                      },
-                                      "type": "Text",
-                                    },
-                                  },
-                                  "type": "Tooltip",
-                                },
-                              ],
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "Test justification text that is longer than twenty characters",
-                                  },
-                                  "type": "Text",
-                                },
-                                null,
-                              ],
-                              "direction": "vertical",
-                            },
-                            "type": "Box",
-                          },
-                        ],
-                        "direction": "vertical",
-                      },
-                      "type": "Box",
-                    },
-                  },
-                  "type": "Section",
-                },
-                {
-                  "key": null,
-                  "props": {
-                    "children": [
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Request from",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The site requesting the permission",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "alignment": "end",
-                                      "children": "https://example.com",
-                                    },
-                                    "type": "Text",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Recipient",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The site requesting the permission",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": {
-                                        "key": null,
-                                        "props": {
-                                          "children": "0x12345...67890",
-                                        },
-                                        "type": "Text",
-                                      },
-                                      "content": "0x1234567890123456789012345678901234567890",
-                                    },
-                                    "type": "Tooltip",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Network",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The network on which the permission is being requested",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "alignment": "end",
-                                      "children": "Ethereum Mainnet",
-                                    },
-                                    "type": "Text",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Token",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The token being requested",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": {
-                                        "key": null,
-                                        "props": {
-                                          "children": "ETH",
-                                          "href": "https://etherscan.io/address/0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3",
-                                        },
-                                        "type": "Link",
-                                      },
-                                      "content": "0x38c4A...611F3",
-                                    },
-                                    "type": "Tooltip",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      null,
-                      null,
-                    ],
-                  },
-                  "type": "Section",
-                },
-                undefined,
-              ],
-              "direction": "vertical",
-            },
-            "type": "Box",
-          },
-        },
-        "type": "Box",
-      },
-      {
-        "key": null,
-        "props": {
-          "children": [
-            {
-              "key": null,
-              "props": {
-                "children": "Cancel",
-                "name": "cancel-button",
-                "variant": "destructive",
-              },
-              "type": "Button",
-            },
-            {
-              "key": null,
-              "props": {
-                "children": "Grant",
-                "disabled": false,
-                "name": "grant-button",
-                "variant": "primary",
-              },
-              "type": "Button",
-            },
-          ],
-        },
-        "type": "Footer",
-      },
-    ],
-  },
-  "type": "Container",
-}
-`);
-
-        // resolve the fiat balance, which triggers a re-render
-        resolveFiatBalancePromise();
-
-        // allow the event loop to run the re-render
-        await new Promise((resolve) => setTimeout(resolve, 0));
-
-        const confirmationContentWithFiatBalance =
-          await lifecycleHandlers.createConfirmationContent({
-            context: mockContext,
-            metadata: mockMetadata,
-            origin: mockOrigin,
-            chainId: 1,
-            scanDappUrlResult: null,
-            scanAddressResult: null,
-            existingPermissionsStatus: ExistingPermissionsState.None,
-            isGrantDisabled: false,
-          });
-
-        // concrete token balance, concrete fiat balance
-        expect(confirmationContentWithFiatBalance).toMatchInlineSnapshot(`
-{
-  "key": null,
-  "props": {
-    "children": [
-      {
-        "key": null,
-        "props": {
-          "children": {
-            "key": null,
-            "props": {
-              "children": [
-                {
-                  "key": null,
-                  "props": {
-                    "center": true,
-                    "children": [
-                      {
-                        "key": null,
-                        "props": {
-                          "children": "Permission request",
-                          "size": "lg",
-                        },
-                        "type": "Heading",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "children": "This site wants permissions to spend your tokens.",
-                        },
-                        "type": "Text",
-                      },
-                    ],
-                  },
-                  "type": "Box",
-                },
-                {
-                  "key": null,
-                  "props": {
-                    "children": {
-                      "key": null,
-                      "props": {
-                        "children": [
-                          {
-                            "key": null,
-                            "props": {
-                              "alignment": "space-between",
-                              "children": {
-                                "key": null,
-                                "props": {
-                                  "children": [
-                                    {
-                                      "key": null,
-                                      "props": {
-                                        "children": "Account",
-                                      },
-                                      "type": "Text",
-                                    },
-                                    {
-                                      "key": null,
-                                      "props": {
-                                        "children": {
-                                          "key": null,
-                                          "props": {
-                                            "color": "muted",
-                                            "name": "question",
-                                            "size": "inherit",
-                                          },
-                                          "type": "Icon",
-                                        },
-                                        "content": {
-                                          "key": null,
-                                          "props": {
-                                            "children": "The account from which the permission is being granted.",
-                                          },
-                                          "type": "Text",
-                                        },
-                                      },
-                                      "type": "Tooltip",
-                                    },
-                                  ],
-                                  "direction": "horizontal",
-                                },
-                                "type": "Box",
-                              },
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "chainIds": [
-                                "eip155:1",
-                              ],
-                              "name": "account-selector",
-                              "switchGlobalAccount": false,
-                              "value": "eip155:1:0x1234567890123456789012345678901234567890",
-                            },
-                            "type": "AccountSelector",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "children": "This account will be upgraded to a smart account to complete this permission.",
-                              "color": "warning",
-                              "size": "sm",
-                            },
-                            "type": "Text",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "alignment": "end",
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "$1000",
-                                  },
-                                  "type": "Text",
-                                },
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": [
-                                      "1",
-                                      " ",
-                                      "available",
-                                    ],
-                                  },
-                                  "type": "Text",
-                                },
-                              ],
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                        ],
-                        "direction": "vertical",
-                      },
-                      "type": "Box",
-                    },
-                  },
-                  "type": "Section",
-                },
-                false,
-                false,
-                {
-                  "key": null,
-                  "props": {
-                    "children": {
-                      "key": null,
-                      "props": {
-                        "alignment": "space-between",
-                        "children": [
-                          {
-                            "key": null,
-                            "props": {
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "Justification",
-                                  },
-                                  "type": "Text",
-                                },
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": {
-                                      "key": null,
-                                      "props": {
-                                        "color": "muted",
-                                        "name": "question",
-                                        "size": "inherit",
-                                      },
-                                      "type": "Icon",
-                                    },
-                                    "content": {
-                                      "key": null,
-                                      "props": {
-                                        "children": "Justification given by the recipient for requesting this permission.",
-                                      },
-                                      "type": "Text",
-                                    },
-                                  },
-                                  "type": "Tooltip",
-                                },
-                              ],
-                              "direction": "horizontal",
-                            },
-                            "type": "Box",
-                          },
-                          {
-                            "key": null,
-                            "props": {
-                              "children": [
-                                {
-                                  "key": null,
-                                  "props": {
-                                    "children": "Test justification text that is longer than twenty characters",
-                                  },
-                                  "type": "Text",
-                                },
-                                null,
-                              ],
-                              "direction": "vertical",
-                            },
-                            "type": "Box",
-                          },
-                        ],
-                        "direction": "vertical",
-                      },
-                      "type": "Box",
-                    },
-                  },
-                  "type": "Section",
-                },
-                {
-                  "key": null,
-                  "props": {
-                    "children": [
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Request from",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The site requesting the permission",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "alignment": "end",
-                                      "children": "https://example.com",
-                                    },
-                                    "type": "Text",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Recipient",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The site requesting the permission",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": {
-                                        "key": null,
-                                        "props": {
-                                          "children": "0x12345...67890",
-                                        },
-                                        "type": "Text",
-                                      },
-                                      "content": "0x1234567890123456789012345678901234567890",
-                                    },
-                                    "type": "Tooltip",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Network",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The network on which the permission is being requested",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "alignment": "end",
-                                      "children": "Ethereum Mainnet",
-                                    },
-                                    "type": "Text",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      {
-                        "key": null,
-                        "props": {
-                          "alignment": "space-between",
-                          "children": [
-                            {
-                              "key": null,
-                              "props": {
-                                "alignment": "space-between",
-                                "children": [
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": [
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": "Token",
-                                          },
-                                          "type": "Text",
-                                        },
-                                        {
-                                          "key": null,
-                                          "props": {
-                                            "children": {
-                                              "key": null,
-                                              "props": {
-                                                "color": "muted",
-                                                "name": "question",
-                                                "size": "inherit",
-                                              },
-                                              "type": "Icon",
-                                            },
-                                            "content": {
-                                              "key": null,
-                                              "props": {
-                                                "children": "The token being requested",
-                                              },
-                                              "type": "Text",
-                                            },
-                                          },
-                                          "type": "Tooltip",
-                                        },
-                                      ],
-                                      "direction": "horizontal",
-                                    },
-                                    "type": "Box",
-                                  },
-                                  null,
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                            {
-                              "key": null,
-                              "props": {
-                                "children": [
-                                  null,
-                                  {
-                                    "key": null,
-                                    "props": {
-                                      "children": {
-                                        "key": null,
-                                        "props": {
-                                          "children": "ETH",
-                                          "href": "https://etherscan.io/address/0x38c4A4F071d33d6Cf83e2e81F12D9B5D30E611F3",
-                                        },
-                                        "type": "Link",
-                                      },
-                                      "content": "0x38c4A...611F3",
-                                    },
-                                    "type": "Tooltip",
-                                  },
-                                ],
-                                "direction": "horizontal",
-                              },
-                              "type": "Box",
-                            },
-                          ],
-                          "direction": "horizontal",
-                        },
-                        "type": "Box",
-                      },
-                      null,
-                      null,
-                    ],
-                  },
-                  "type": "Section",
-                },
-                undefined,
-              ],
-              "direction": "vertical",
-            },
-            "type": "Box",
-          },
-        },
-        "type": "Box",
-      },
-      {
-        "key": null,
-        "props": {
-          "children": [
-            {
-              "key": null,
-              "props": {
-                "children": "Cancel",
-                "name": "cancel-button",
-                "variant": "destructive",
-              },
-              "type": "Button",
-            },
-            {
-              "key": null,
-              "props": {
-                "children": "Grant",
-                "disabled": false,
-                "name": "grant-button",
-                "variant": "primary",
-              },
-              "type": "Button",
-            },
-          ],
-        },
-        "type": "Footer",
-      },
-    ],
-  },
-  "type": "Container",
-}
-`);
-      });
-
-      it('cancels the balance loading when the account is changed', async () => {
-        const {
-          permissionHandler,
-          getLifecycleHandlers,
-          getBoundEvent,
-          tokenMetadataService,
-          updateContext,
-        } = setupTest();
-
-        let resolveTokenBalancePromise: () => void = (): void => {
-          throw new Error('Function should never be called');
-        };
-        const tokenBalancePromise = new Promise<TokenBalanceAndMetadata>(
-          (resolve) => {
-            resolveTokenBalancePromise = (): void =>
-              resolve(mockTokenBalanceAndMetadata);
-          },
-        );
-
-        tokenMetadataService.getTokenBalanceAndMetadata.mockReturnValueOnce(
-          tokenBalancePromise,
-        );
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        lifecycleHandlers.onConfirmationCreated?.({
-          interfaceId: mockInterfaceId,
-          initialContext: mockContext,
-          updateContext,
-        });
-
-        const accountSelectorChangeHandler = getBoundEvent({
-          elementName: 'account-selector',
-          eventType: 'InputChangeEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(accountSelectorChangeHandler).toBeDefined();
-
-        const mockAddress2Caip10 = `eip155:1:${mockAddress2}`;
-
-        await accountSelectorChangeHandler?.({
-          event: {
-            value: { addresses: [mockAddress2Caip10] } as any,
-            name: 'account-selector',
-            type: UserInputEventType.InputChangeEvent,
-          },
-          interfaceId: mockInterfaceId,
-        });
-
-        resolveTokenBalancePromise();
-        await new Promise((resolve) => setTimeout(resolve, 0));
-
-        // update context is called 4 times:
-        // 1. After the account is changed
-        // 2. After the account upgrade status is fetched for the new account
-        // 3. After the token balance is resolved for the second account
-        // 4. After the fiat balance is resolved for the second account
-        // if we didn't cancel the original balance loading, there would be another 2 instances
-        // 5. After the original token balance is resolved for the first account
-        // 6. After the original fiat balance is resolved for the first account
-        expect(updateContext).toHaveBeenCalledTimes(4);
-      });
-
-      it('unbinds the event handlers when the confirmation is resolved', async () => {
-        const {
-          permissionHandler,
-          getLifecycleHandlers,
-          getUnboundEvent,
-          getBoundEvent,
-          updateContext,
-        } = setupTest();
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const { onConfirmationResolved, onConfirmationCreated } =
-          getLifecycleHandlers();
-
-        onConfirmationCreated?.({
-          interfaceId: mockInterfaceId,
-          initialContext: mockContext,
-          updateContext,
-        });
-
-        const accountSelectorBoundEvent = getBoundEvent({
-          elementName: 'account-selector',
-          eventType: 'InputChangeEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        const showMoreButtonBoundEvent = getBoundEvent({
-          elementName: 'show-more-justification',
-          eventType: 'ButtonClickEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(accountSelectorBoundEvent).toBeDefined();
-        expect(showMoreButtonBoundEvent).toBeDefined();
-
-        onConfirmationResolved?.();
-
-        const accountSelectorUnboundEvent = getUnboundEvent({
-          elementName: 'account-selector',
-          eventType: 'InputChangeEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        const showMoreButtonUnboundEvent = getUnboundEvent({
-          elementName: 'show-more-justification',
-          eventType: 'ButtonClickEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(accountSelectorUnboundEvent).toBeDefined();
-        expect(showMoreButtonUnboundEvent).toBeDefined();
-      });
-
-      it('binds rule handlers even when isAdjustmentAllowed is false', async () => {
-        const rule: RuleDefinition<TestContextType, TestMetadataType> = {
-          name: 'amountPerSecond',
-          label: 'amountLabel',
-          type: 'number',
-          getRuleData: () => ({
-            value: '0x1',
-            isVisible: true,
-            isEditable: false,
-          }),
-          updateContext: (context) => context,
-        };
-        const {
-          permissionHandler,
-          getLifecycleHandlers,
-          getBoundEvent,
-          updateContext,
-        } = setupTest({ rules: [rule] });
-
-        await permissionHandler.handlePermissionRequest(mockOrigin);
-
-        const lifecycleHandlers = getLifecycleHandlers();
-
-        lifecycleHandlers.onConfirmationCreated?.({
-          interfaceId: mockInterfaceId,
-          initialContext: mockContext,
-          updateContext,
-        });
-
-        // Try to get a rule input handler - it should still be bound
-        const ruleInputHandler = getBoundEvent({
-          elementName: 'amountPerSecond', // Example rule input field
-          eventType: 'InputChangeEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(ruleInputHandler).toBeDefined();
-
-        // Account selector should still be bound (it's allowed even when adjustment is not allowed)
-        const accountSelectorHandler = getBoundEvent({
-          elementName: 'account-selector',
-          eventType: 'InputChangeEvent',
-          interfaceId: mockInterfaceId,
-        });
-
-        expect(accountSelectorHandler).toBeDefined();
+      expect(dependencies.createConfirmationContent).toHaveBeenCalledWith({
+        context: mockContext,
+        metadata: mockMetadata,
       });
     });
   });
